feat(templatecenter): add daemonless template image export and artifact management
Introduce template image export via skopeo/umoci, rootfs artifact
build/cleanup, and file-level fingerprint deduplication.

- Add image export pipeline (skopeo copy + umoci unpack + ext4 mkfs)
- Add artifact build, cleanup, and distribution orchestration
- Add file-level content fingerprint for artifact deduplication
- Add redo support with resume-from-phase optimization
- Add job runner, DTO, repository, and constants for image jobs
- Add template request validation and job lifecycle management
- Update copyright headers to 2026
