+++ CubeMaster/pkg/templatecenter/template_image.go
@@ -1,153 +1,23 @@
-// Copyright (c) 2024 Tencent Inc.
+// Copyright (c) 2026 Tencent Inc.
 // SPDX-License-Identifier: Apache-2.0
 //
 
 package templatecenter
 
 import (
-	"bytes"
 	"context"
-	"crypto/sha256"
-	"encoding/base64"
-	"encoding/hex"
-	"encoding/json"
 	"errors"
 	"fmt"
-	"io"
-	"net/url"
-	"os"
-	"os/exec"
-	"path/filepath"
-	"regexp"
-	"sort"
-	"strconv"
-	"strings"
-	"sync"
-	"syscall"
-	"time"
-
 	"github.com/google/uuid"
-	cubeboxv1 "github.com/tencentcloud/CubeSandbox/CubeMaster/api/services/cubebox/v1"
-	imagev1 "github.com/tencentcloud/CubeSandbox/CubeMaster/api/services/images/v1"
-	"github.com/tencentcloud/CubeSandbox/CubeMaster/pkg/base/config"
 	"github.com/tencentcloud/CubeSandbox/CubeMaster/pkg/base/constants"
 	"github.com/tencentcloud/CubeSandbox/CubeMaster/pkg/base/db/models"
 	"github.com/tencentcloud/CubeSandbox/CubeMaster/pkg/base/log"
 	"github.com/tencentcloud/CubeSandbox/CubeMaster/pkg/base/node"
-	"github.com/tencentcloud/CubeSandbox/CubeMaster/pkg/cubelet"
-	"github.com/tencentcloud/CubeSandbox/CubeMaster/pkg/errorcode"
 	"github.com/tencentcloud/CubeSandbox/CubeMaster/pkg/service/sandbox/types"
-	CubeLog "github.com/tencentcloud/CubeSandbox/cubelog"
-	"golang.org/x/sys/unix"
 	"gorm.io/gorm"
+	"os"
+	"strings"
 )
-
-const (
-	ArtifactStatusPending  = "PENDING"
-	ArtifactStatusBuilding = "BUILDING"
-	ArtifactStatusReady    = "READY"
-	ArtifactStatusFailed   = "FAILED"
-
-	JobStatusPending = "PENDING"
-	JobStatusRunning = "RUNNING"
-	JobStatusReady   = "READY"
-	JobStatusFailed  = "FAILED"
-
-	JobOperationCreate           = "CREATE"
-	JobOperationRedo             = "REDO"
-	JobOperationCommit           = "COMMIT"
-	JobOperationLegacy           = "LEGACY"
-	JobOperationSnapshotCreate   = "SNAPSHOT_CREATE"
-	JobOperationSnapshotRollback = "SNAPSHOT_ROLLBACK"
-	JobOperationSnapshotDelete   = "SNAPSHOT_DELETE"
-
-	JobResourceTypeSnapshot = "snapshot"
-	JobResourceTypeTemplate = "template"
-
-	RedoModeAll         = "ALL"
-	RedoModeNodes       = "NODES"
-	RedoModeFailedOnly  = "FAILED_ONLY"
-	RedoModeFailedNodes = "FAILED_NODES"
-
-	JobPhasePulling            = "PULLING"
-	JobPhaseUnpacking          = "UNPACKING"
-	JobPhaseBuildingExt4       = "BUILDING_EXT4"
-	JobPhaseGeneratingJSON     = "GENERATING_JSON"
-	JobPhaseDistributing       = "DISTRIBUTING"
-	JobPhaseCreatingTemplate   = "CREATING_TEMPLATE"
-	JobPhaseSnapshotting       = "SNAPSHOTTING"
-	JobPhaseRegistering        = "REGISTERING"
-	JobPhaseRollbackPreparing  = "ROLLBACK_PREPARING"
-	JobPhaseRollbackDriving    = "ROLLBACK_DRIVING"
-	JobPhaseRollbackRecovering = "ROLLBACK_RECOVERING"
-	JobPhaseDeleting           = "DELETING"
-	JobPhaseReady              = "READY"
-
-	defaultTemplateCPU         = "2000m"
-	defaultTemplateMemory      = "2000Mi"
-	defaultTemplateArtifactTTL = 7 * 24 * time.Hour
-	defaultArtifactStoreDir    = "/data/CubeMaster/storage"
-	fallbackArtifactStoreDir   = "cubemaster-rootfs-artifacts-store"
-	rootfsWritableVolumeName   = "cube_rootfs_rw"
-	defaultDistributionWorkers = 4
-	legacyRequestIDPrefix      = "legacy-"
-)
-
-var getTemplateImageConfig = config.GetConfig
-var deleteRootfsArtifactRecord = func(ctx context.Context, artifactID string) error {
-	return store.db.WithContext(ctx).Unscoped().Table(constants.RootfsArtifactTableName).
-		Where("artifact_id = ?", artifactID).Delete(&models.RootfsArtifact{}).Error
-}
-var executableLookPath = exec.LookPath
-
-var ErrNoFailedTemplateReplicas = errors.New("no failed template replicas matched redo request")
-
-type dockerInspectImage struct {
-	ID          string            `json:"Id"`
-	RepoDigests []string          `json:"RepoDigests"`
-	Config      dockerImageConfig `json:"Config"`
-}
-
-type skopeoInspectImage struct {
-	Name       string               `json:"Name"`
-	Digest     string               `json:"Digest"`
-	LayersData []skopeoInspectLayer `json:"LayersData"`
-}
-
-// skopeoInspectLayer mirrors a single entry of the LayersData array returned by
-// `skopeo inspect`. Size is the compressed (on-registry) size of the layer blob
-// in bytes.
-type skopeoInspectLayer struct {
-	Size int64 `json:"Size"`
-}
-
-type skopeoInspectConfig struct {
-	Config dockerImageConfig `json:"config"`
-}
-
-type dockerImageConfig struct {
-	Entrypoint []string `json:"Entrypoint"`
-	Cmd        []string `json:"Cmd"`
-	Env        []string `json:"Env"`
-	WorkingDir string   `json:"WorkingDir"`
-	User       string   `json:"User"`
-}
-
-type resolvedSourceImage struct {
-	localRef       string
-	digest         string
-	config         dockerImageConfig
-	configJSON     string
-	masterNodeIP   string
-	useDockerless  bool
-	skopeoAuthFile string
-	// compressedSizeBytes is the sum of the compressed layer blob sizes reported
-	// by `skopeo inspect` (LayersData[].Size). It is only populated on the
-	// dockerless path and lets the disk-space pre-check estimate the image size
-	// without invoking the docker daemon. Zero means "unknown".
-	compressedSizeBytes int64
-	cleanup             func(context.Context)
-}
 
 func nextAttemptNoFromLatest(latestAttemptNo int32) int32 {
 	if latestAttemptNo <= 0 {
@@ -377,350 +247,6 @@ func GetRootfsArtifactInfo(ctx context.Context, artifactID string) (*types.Rootf
 	return artifactModelToInfo(record), nil
 }
 
-func normalizeRedoTemplateImageRequest(req *types.RedoTemplateFromImageReq) (*types.RedoTemplateFromImageReq, error) {
-	if req == nil {
-		return nil, errors.New("request is nil")
-	}
-	if req.Request == nil || strings.TrimSpace(req.RequestID) == "" {
-		return nil, errors.New("requestID is required")
-	}
-	if strings.TrimSpace(req.TemplateID) == "" {
-		return nil, errors.New("template_id is required")
-	}
-	cloned := *req
-	if len(req.DistributionScope) > 0 {
-		cloned.DistributionScope = append([]string(nil), req.DistributionScope...)
-	}
-	return &cloned, nil
-}
-
-func allowRedoResumePhase(job *models.TemplateImageJob) error {
-	if job == nil {
-		return ErrTemplateNotFound
-	}
-	switch strings.ToUpper(strings.TrimSpace(job.Phase)) {
-	case "", JobPhasePulling:
-		return errors.New("template redo is not allowed before source image has been pulled successfully")
-	default:
-		return nil
-	}
-}
-
-func getTemplateImageJobRecordByID(ctx context.Context, jobID string) (*models.TemplateImageJob, error) {
-	record := &models.TemplateImageJob{}
-	if err := store.db.WithContext(ctx).Table(constants.TemplateImageJobTableName).
-		Where("job_id = ?", jobID).First(record).Error; err != nil {
-		return nil, err
-	}
-	return record, nil
-}
-
-func unmarshalTemplateImageJobRequest(payload string) (*types.CreateTemplateFromImageReq, error) {
-	req := &types.CreateTemplateFromImageReq{}
-	if err := json.Unmarshal([]byte(payload), req); err != nil {
-		return nil, err
-	}
-	req.Request = &types.Request{RequestID: uuid.NewString()}
-	return normalizeTemplateImageRequest(req)
-}
-
-func determineRedoMode(req *types.RedoTemplateFromImageReq) string {
-	switch {
-	case req == nil:
-		return RedoModeAll
-	case req.FailedOnly && len(req.DistributionScope) > 0:
-		return RedoModeFailedNodes
-	case req.FailedOnly:
-		return RedoModeFailedOnly
-	case len(req.DistributionScope) > 0:
-		return RedoModeNodes
-	default:
-		return RedoModeAll
-	}
-}
-
-func replicaNeedsRedo(replica models.TemplateReplica) bool {
-	return replica.Status != ReplicaStatusReady || replica.CleanupRequired
-}
-
-func failedRedoScope(replicas []models.TemplateReplica) []string {
-	failedScope := make([]string, 0, len(replicas))
-	for _, replica := range replicas {
-		if !replicaNeedsRedo(replica) {
-			continue
-		}
-		if replica.NodeID != "" {
-			failedScope = append(failedScope, replica.NodeID)
-			continue
-		}
-		if replica.NodeIP != "" {
-			failedScope = append(failedScope, replica.NodeIP)
-		}
-	}
-	return failedScope
-}
-
-func marshalRedoScope(scope []string) string {
-	if len(scope) == 0 {
-		return ""
-	}
-	payload, err := json.Marshal(scope)
-	if err != nil {
-		return ""
-	}
-	return string(payload)
-}
-
-func unmarshalRedoScope(scopeJSON string) []string {
-	if strings.TrimSpace(scopeJSON) == "" {
-		return nil
-	}
-	var scope []string
-	if err := json.Unmarshal([]byte(scopeJSON), &scope); err != nil {
-		return nil
-	}
-	return scope
-}
-
-func determineRedoResumePhase(job *models.TemplateImageJob, replicas []models.TemplateReplica) string {
-	if job != nil {
-		switch strings.ToUpper(job.Phase) {
-		case JobPhasePulling, JobPhaseUnpacking, JobPhaseBuildingExt4, JobPhaseGeneratingJSON:
-			return JobPhaseBuildingExt4
-		case JobPhaseDistributing:
-			return JobPhaseDistributing
-		case JobPhaseCreatingTemplate, JobPhaseSnapshotting, JobPhaseRegistering:
-			return JobPhaseSnapshotting
-		}
-	}
-	for _, replica := range replicas {
-		if replica.Status == ReplicaStatusReady {
-			continue
-		}
-		switch strings.ToUpper(replica.LastErrorPhase) {
-		case ReplicaPhaseDistributing:
-			return JobPhaseDistributing
-		case ReplicaPhaseSnapshotting, ReplicaPhaseFailed:
-			return JobPhaseSnapshotting
-		}
-	}
-	return JobPhaseSnapshotting
-}
-
-func resolveTemplateNodes(instanceType string, scope []string) ([]*node.Node, error) {
-	nodes := healthyTemplateNodes(instanceType)
-	if len(nodes) == 0 {
-		return nil, ErrNoTemplateNodes
-	}
-	if len(scope) == 0 {
-		return nodes, nil
-	}
-	allowed := make(map[string]struct{}, len(scope))
-	for _, item := range scope {
-		item = strings.TrimSpace(item)
-		if item == "" {
-			continue
-		}
-		allowed[item] = struct{}{}
-	}
-	selected := make([]*node.Node, 0, len(nodes))
-	matched := make(map[string]struct{})
-	for _, item := range nodes {
-		if item == nil {
-			continue
-		}
-		if _, ok := allowed[item.ID()]; ok {
-			selected = append(selected, item)
-			matched[item.ID()] = struct{}{}
-			continue
-		}
-		if _, ok := allowed[item.HostIP()]; ok {
-			selected = append(selected, item)
-			matched[item.HostIP()] = struct{}{}
-		}
-	}
-	missing := make([]string, 0)
-	for _, item := range scope {
-		if _, ok := matched[item]; ok {
-			continue
-		}
-		missing = append(missing, item)
-	}
-	if len(missing) > 0 {
-		sort.Strings(missing)
-		return nil, fmt.Errorf("target nodes are not healthy or not found: %s", strings.Join(missing, ","))
-	}
-	if len(selected) == 0 {
-		return nil, ErrNoTemplateNodes
-	}
-	return selected, nil
-}
-
-func resolveRedoTargets(instanceType string, req *types.RedoTemplateFromImageReq, replicas []models.TemplateReplica) ([]*node.Node, error) {
-	if req == nil {
-		return resolveTemplateNodes(instanceType, nil)
-	}
-	baseScope := req.DistributionScope
-	if len(baseScope) == 0 {
-		baseScope = nil
-	}
-	targets, err := resolveTemplateNodes(instanceType, baseScope)
-	if err != nil {
-		return nil, err
-	}
-	if !req.FailedOnly {
-		return targets, nil
-	}
-	failedScope := failedRedoScope(replicas)
-	if len(failedScope) == 0 {
-		return nil, ErrNoFailedTemplateReplicas
-	}
-	failedSet := make(map[string]struct{}, len(failedScope))
-	for _, item := range failedScope {
-		if strings.TrimSpace(item) == "" {
-			continue
-		}
-		failedSet[item] = struct{}{}
-	}
-	filtered := make([]*node.Node, 0, len(targets))
-	for _, target := range targets {
-		if target == nil {
-			continue
-		}
-		if _, ok := failedSet[target.ID()]; ok {
-			filtered = append(filtered, target)
-			continue
-		}
-		if _, ok := failedSet[target.HostIP()]; ok {
-			filtered = append(filtered, target)
-		}
-	}
-	if len(filtered) == 0 {
-		return nil, ErrNoFailedTemplateReplicas
-	}
-	return filtered, nil
-}
-
-func prepareLocalSourceImage(ctx context.Context, req *types.CreateTemplateFromImageReq, downloadBaseURL string) (*resolvedSourceImage, error) {
-	if req == nil {
-		return nil, errors.New("request is nil")
-	}
-	// In dockerless mode there is no local docker daemon to hold the image, so a
-	// redo re-resolves the source from the registry via skopeo. This intentionally
-	// relaxes the docker-path requirement that the image still exist locally.
-	if hasDockerlessRootfsExportTools() {
-		return prepareDockerlessSourceImage(ctx, req, downloadBaseURL)
-	}
-	inspectOutput, err := dockerOutput(ctx, "", "image", "inspect", "--", req.SourceImageRef)
-	if err != nil {
-		return nil, fmt.Errorf("redo requires source image %s to still exist locally: %w", req.SourceImageRef, err)
-	}
-	var inspectList []dockerInspectImage
-	if err := json.Unmarshal(inspectOutput, &inspectList); err != nil {
-		return nil, fmt.Errorf("unmarshal local docker inspect output: %w", err)
-	}
-	if len(inspectList) == 0 {
-		return nil, fmt.Errorf("docker image inspect returned empty result for %s", req.SourceImageRef)
-	}
-	inspectInfo := inspectList[0]
-	configJSON, err := json.Marshal(inspectInfo.Config)
-	if err != nil {
-		return nil, fmt.Errorf("marshal image config: %w", err)
-	}
-	return &resolvedSourceImage{
-		localRef:     req.SourceImageRef,
-		digest:       firstNonEmptyDigest(inspectInfo),
-		config:       inspectInfo.Config,
-		configJSON:   string(configJSON),
-		masterNodeIP: normalizeBaseURL(downloadBaseURL),
-	}, nil
-}
-
-func buildReplicaForDistribution(target *node.Node, req *types.CreateCubeSandboxReq, artifactID, jobID string) ReplicaStatus {
-	spec := ""
-	instanceType := ""
-	if req != nil {
-		spec = calculateRequestSpec(req)
-		instanceType = req.InstanceType
-	}
-	return ReplicaStatus{
-		NodeID:          target.ID(),
-		NodeIP:          target.HostIP(),
-		InstanceType:    instanceType,
-		Spec:            spec,
-		Status:          ReplicaStatusFailed,
-		Phase:           ReplicaPhaseDistributing,
-		ArtifactID:      artifactID,
-		LastJobID:       jobID,
-		LastErrorPhase:  ReplicaPhaseDistributing,
-		CleanupRequired: true,
-	}
-}
-
-func cleanupArtifactOnNodes(ctx context.Context, artifactID string, targets []*node.Node) error {
-	if artifactID == "" {
-		return nil
-	}
-	var cleanupErr error
-	for _, target := range targets {
-		if target == nil {
-			continue
-		}
-		rsp, err := deleteImageOnCubelet(ctx, getCubeletAddrForDelete(target.HostIP()), &imagev1.DestroyImageRequest{
-			RequestID: uuid.NewString(),
-			Spec: &imagev1.ImageSpec{
-				Image: artifactID,
-			},
-		})
-		if err != nil {
-			if isIgnorableArtifactDeleteError(err) {
-				continue
-			}
-			cleanupErr = errors.Join(cleanupErr, fmt.Errorf("delete artifact %s on node %s: %w", artifactID, target.ID(), err))
-			continue
-		}
-		if rsp.GetRet() != nil && int(rsp.GetRet().GetRetCode()) != int(errorcode.ErrorCode_Success) {
-			if isIgnorableArtifactDeleteMessage(rsp.GetRet().GetRetMsg()) {
-				continue
-			}
-			cleanupErr = errors.Join(cleanupErr, fmt.Errorf("delete artifact %s on node %s failed: %s", artifactID, target.ID(), rsp.GetRet().GetRetMsg()))
-		}
-	}
-	return cleanupErr
-}
-
-func cleanupTemplateReplicasOnNodes(ctx context.Context, templateID string, replicas []models.TemplateReplica, targets []*node.Node) error {
-	if len(replicas) == 0 || len(targets) == 0 {
-		return nil
-	}
-	allowed := make(map[string]struct{}, len(targets)*2)
-	for _, target := range targets {
-		if target == nil {
-			continue
-		}
-		allowed[target.ID()] = struct{}{}
-		if target.HostIP() != "" {
-			allowed[target.HostIP()] = struct{}{}
-		}
-	}
-	locators := make([]templateCleanupLocator, 0, len(replicas))
-	for _, replica := range replicas {
-		if _, ok := allowed[replica.NodeID]; !ok {
-			if _, ok := allowed[replica.NodeIP]; !ok {
-				continue
-			}
-		}
-		locators = append(locators, templateCleanupLocator{
-			NodeID: replica.NodeID,
-			NodeIP: replica.NodeIP,
-		})
-	}
-	if len(locators) == 0 {
-		return nil
-	}
-	return cleanupTemplateReplicasWithLocators(ctx, templateID, locators)
-}
-
 func OpenRootfsArtifact(ctx context.Context, artifactID, token string) (*models.RootfsArtifact, *os.File, error) {
 	record, err := getRootfsArtifactByID(ctx, artifactID)
 	if err != nil {
@@ -737,2418 +263,4 @@ func OpenRootfsArtifact(ctx context.Context, artifactID, token string) (*models.
 		return nil, nil, err
 	}
 	return record, f, nil
-}
-
-func runTemplateImageJob(ctx context.Context, jobID string, req *types.CreateTemplateFromImageReq, downloadBaseURL string) {
-	logger := log.G(ctx).WithFields(map[string]any{
-		"job_id":      jobID,
-		"template_id": req.TemplateID,
-		"image":       req.SourceImageRef,
-	})
-	if err := updateTemplateImageJob(ctx, jobID, map[string]any{
-		"status":   JobStatusRunning,
-		"phase":    JobPhasePulling,
-		"progress": 5,
-	}); err != nil {
-		logger.Errorf("update job start fail: %v", err)
-		return
-	}
-	if err := ensureArtifactBuildPreflight(ctx); err != nil {
-		_ = updateTemplateImageJob(ctx, jobID, map[string]any{
-			"status":        JobStatusFailed,
-			"phase":         JobPhasePulling,
-			"progress":      100,
-			"error_message": err.Error(),
-		})
-		return
-	}
-	source, err := prepareSourceImage(ctx, req, downloadBaseURL)
-	if err != nil {
-		_ = updateTemplateImageJob(ctx, jobID, map[string]any{
-			"status":        JobStatusFailed,
-			"phase":         JobPhasePulling,
-			"progress":      100,
-			"error_message": err.Error(),
-		})
-		return
-	}
-	if source.cleanup != nil {
-		defer source.cleanup(ctx)
-	}
-	fingerprint := buildTemplateSpecFingerprint(req, source.digest)
-	artifactID := buildArtifactID(fingerprint)
-	if err := updateTemplateImageJob(ctx, jobID, map[string]any{
-		"artifact_id":               artifactID,
-		"template_spec_fingerprint": fingerprint,
-		"source_image_digest":       source.digest,
-		"phase":                     JobPhaseUnpacking,
-		"progress":                  20,
-	}); err != nil {
-		logger.Errorf("update job source metadata fail: %v", err)
-	}
-	artifact, generatedReq, builtFreshArtifact, err := ensureRootfsArtifact(ctx, req, source, downloadBaseURL)
-	if err != nil {
-		_ = updateTemplateImageJob(ctx, jobID, map[string]any{
-			"status":                    JobStatusFailed,
-			"phase":                     JobPhaseBuildingExt4,
-			"artifact_id":               artifactID,
-			"template_spec_fingerprint": fingerprint,
-			"artifact_status":           ArtifactStatusFailed,
-			"error_message":             err.Error(),
-			"progress":                  100,
-		})
-		return
-	}
-	if err := updateTemplateImageJob(ctx, jobID, map[string]any{
-		"artifact_id":               artifact.ArtifactID,
-		"template_spec_fingerprint": artifact.TemplateSpecFingerprint,
-		"source_image_digest":       artifact.SourceImageDigest,
-		"artifact_status":           artifact.Status,
-		"phase":                     JobPhaseDistributing,
-		"progress":                  70,
-	}); err != nil {
-		logger.Errorf("update job artifact fail: %v", err)
-	}
-	readyTargets, expected, ready, failed, distErr := distributeRootfsArtifact(ctx, req, generatedReq, artifact, req.TemplateID, jobID)
-	if err := updateTemplateImageJob(ctx, jobID, map[string]any{
-		"phase":               JobPhaseCreatingTemplate,
-		"progress":            85,
-		"expected_node_count": expected,
-		"ready_node_count":    ready,
-		"failed_node_count":   failed,
-		"error_message":       errorString(distErr),
-	}); err != nil {
-		logger.Errorf("update distribution status fail: %v", err)
-	}
-	if expected > 0 && ready == 0 {
-		if builtFreshArtifact {
-			if cleanupErr := cleanupFailedRootfsArtifact(ctx, artifact, req.InstanceType); cleanupErr != nil {
-				logger.Errorf("cleanup fresh rootfs artifact after distribution failure fail: %v", cleanupErr)
-			}
-		}
-		_ = updateTemplateImageJob(ctx, jobID, map[string]any{
-			"status":        JobStatusFailed,
-			"phase":         JobPhaseDistributing,
-			"progress":      100,
-			"error_message": fmt.Sprintf("artifact distribution failed on all %d nodes: %v", expected, distErr),
-		})
-		return
-	}
-	var info *TemplateInfo
-	storedReq, err := normalizeStoredTemplateRequest(generatedReq)
-	if err != nil {
-		_ = updateTemplateImageJob(ctx, jobID, map[string]any{
-			"status":          JobStatusFailed,
-			"phase":           JobPhaseCreatingTemplate,
-			"progress":        100,
-			"template_status": StatusFailed,
-			"error_message":   err.Error(),
-		})
-		return
-	}
-	if _, err := ensureTemplateDefinition(ctx, req.TemplateID, storedReq, generatedReq.InstanceType, constants.GetAppSnapshotVersion(generatedReq.Annotations)); err != nil {
-		_ = updateTemplateImageJob(ctx, jobID, map[string]any{
-			"status":          JobStatusFailed,
-			"phase":           JobPhaseCreatingTemplate,
-			"progress":        100,
-			"template_status": StatusFailed,
-			"error_message":   err.Error(),
-		})
-		return
-	}
-	replicas, persistErr := createTemplateReplicasOnNodes(ctx, req.TemplateID, generatedReq, readyTargets, replicaRunOptions{
-		ArtifactID: artifact.ArtifactID,
-		JobID:      jobID,
-	})
-	if persistErr != nil {
-		err = persistErr
-	} else {
-		info, err = finalizeTemplateReplicas(ctx, req.TemplateID, generatedReq.InstanceType, constants.GetAppSnapshotVersion(generatedReq.Annotations), replicas)
-	}
-	if err != nil {
-		if builtFreshArtifact {
-			if cleanupErr := cleanupFailedRootfsArtifact(ctx, artifact, req.InstanceType); cleanupErr != nil {
-				logger.Errorf("cleanup fresh rootfs artifact after create template error fail: %v", cleanupErr)
-			}
-		}
-		_ = updateTemplateImageJob(ctx, jobID, map[string]any{
-			"status":          JobStatusFailed,
-			"phase":           JobPhaseCreatingTemplate,
-			"progress":        100,
-			"template_status": StatusFailed,
-			"error_message":   err.Error(),
-		})
-		return
-	}
-	resultPayload, _ := json.Marshal(info)
-	jobStatus := JobStatusReady
-	jobPhase := JobPhaseReady
-	if info.Status == StatusFailed {
-		if builtFreshArtifact {
-			if cleanupErr := cleanupFailedRootfsArtifact(ctx, artifact, req.InstanceType); cleanupErr != nil {
-				logger.Errorf("cleanup fresh rootfs artifact after failed template status fail: %v", cleanupErr)
-			}
-		}
-		jobStatus = JobStatusFailed
-		jobPhase = JobPhaseCreatingTemplate
-	}
-	_ = updateTemplateImageJob(ctx, jobID, map[string]any{
-		"status":          jobStatus,
-		"phase":           jobPhase,
-		"progress":        100,
-		"template_status": info.Status,
-		"result_json":     string(resultPayload),
-		"error_message":   info.LastError,
-	})
-}
-
-func failRedoTemplateImageJob(ctx context.Context, jobID, phase, message string) {
-	_ = updateTemplateImageJob(ctx, jobID, map[string]any{
-		"status":        JobStatusFailed,
-		"phase":         phase,
-		"progress":      100,
-		"error_message": message,
-	})
-}
-
-func runRedoTemplateImageJob(ctx context.Context, jobID string, req *types.RedoTemplateFromImageReq, downloadBaseURL string) {
-	logger := log.G(ctx).WithFields(map[string]any{
-		"job_id":      jobID,
-		"template_id": req.TemplateID,
-	})
-	jobRecord, err := getTemplateImageJobRecordByID(ctx, jobID)
-	if err != nil {
-		logger.Errorf("lookup redo job fail: %v", err)
-		return
-	}
-	if err := updateTemplateImageJob(ctx, jobID, map[string]any{
-		"status":   JobStatusRunning,
-		"phase":    jobRecord.ResumePhase,
-		"progress": 5,
-	}); err != nil {
-		logger.Errorf("update redo job start fail: %v", err)
-		return
-	}
-	sourceReq, err := unmarshalTemplateImageJobRequest(jobRecord.RequestJSON)
-	if err != nil {
-		failRedoTemplateImageJob(ctx, jobID, jobRecord.ResumePhase, err.Error())
-		return
-	}
-	existingReplicas, err := ListReplicas(ctx, req.TemplateID)
-	if err != nil {
-		failRedoTemplateImageJob(ctx, jobID, jobRecord.ResumePhase, err.Error())
-		return
-	}
-	targets, err := resolveRedoTargets(sourceReq.InstanceType, req, existingReplicas)
-	if err != nil {
-		failRedoTemplateImageJob(ctx, jobID, jobRecord.ResumePhase, err.Error())
-		return
-	}
-	workingReq := newRedoWorkingRequest(sourceReq, targets)
-
-	var artifact *models.RootfsArtifact
-	resumePhase := jobRecord.ResumePhase
-	if resumePhase == "" {
-		resumePhase = JobPhaseSnapshotting
-	}
-	if resumePhase == JobPhaseBuildingExt4 {
-		if err := ensureArtifactBuildPreflight(ctx); err != nil {
-			failRedoTemplateImageJob(ctx, jobID, JobPhaseBuildingExt4, err.Error())
-			return
-		}
-		if jobRecord.ArtifactID != "" {
-			if previousArtifact, lookupErr := getRootfsArtifactByID(ctx, jobRecord.ArtifactID); lookupErr == nil {
-				if previousArtifact.Ext4Path != "" {
-					_ = cleanupLocalRootfsArtifact(previousArtifact.ArtifactID, previousArtifact.Ext4Path)
-				}
-				_ = updateRootfsArtifact(ctx, previousArtifact.ArtifactID, map[string]any{
-					"status":     ArtifactStatusFailed,
-					"last_error": "redo requested after artifact build failure",
-				})
-			}
-		}
-		source, prepErr := prepareLocalSourceImage(ctx, &workingReq, downloadBaseURL)
-		if prepErr != nil {
-			failRedoTemplateImageJob(ctx, jobID, JobPhaseBuildingExt4, prepErr.Error())
-			return
-		}
-		if source.cleanup != nil {
-			defer source.cleanup(ctx)
-		}
-		var generatedReq *types.CreateCubeSandboxReq
-		var builtFresh bool
-		artifact, generatedReq, builtFresh, err = ensureRootfsArtifact(ctx, &workingReq, source, downloadBaseURL)
-		if err != nil {
-			_ = updateTemplateImageJob(ctx, jobID, map[string]any{
-				"status":          JobStatusFailed,
-				"phase":           JobPhaseBuildingExt4,
-				"progress":        100,
-				"artifact_status": ArtifactStatusFailed,
-				"error_message":   err.Error(),
-			})
-			return
-		}
-		workingReq = newRedoWorkingRequest(sourceReq, targets)
-		_ = generatedReq
-		_ = builtFresh
-		jobRecord.ArtifactID = artifact.ArtifactID
-		if err := updateTemplateImageJob(ctx, jobID, map[string]any{
-			"artifact_id":               artifact.ArtifactID,
-			"template_spec_fingerprint": artifact.TemplateSpecFingerprint,
-			"source_image_digest":       artifact.SourceImageDigest,
-			"artifact_status":           artifact.Status,
-			"phase":                     JobPhaseDistributing,
-			"progress":                  60,
-		}); err != nil {
-			logger.Errorf("update redo rebuilt artifact fail: %v", err)
-		}
-		resumePhase = JobPhaseDistributing
-	} else {
-		artifact, err = getRootfsArtifactByID(ctx, jobRecord.ArtifactID)
-		if err != nil {
-			failRedoTemplateImageJob(ctx, jobID, resumePhase, err.Error())
-			return
-		}
-	}
-
-	var generatedReq *types.CreateCubeSandboxReq
-	if strings.TrimSpace(artifact.GeneratedRequestJSON) != "" {
-		generatedReq = &types.CreateCubeSandboxReq{}
-		if err := json.Unmarshal([]byte(artifact.GeneratedRequestJSON), generatedReq); err != nil {
-			generatedReq = nil
-		}
-	}
-	if generatedReq == nil {
-		generatedReq, err = generateTemplateCreateRequest(&workingReq, artifact, dockerImageConfig{}, downloadBaseURL)
-		if artifact.ImageConfigJSON != "" {
-			var imageCfg dockerImageConfig
-			if json.Unmarshal([]byte(artifact.ImageConfigJSON), &imageCfg) == nil {
-				generatedReq, err = generateTemplateCreateRequest(&workingReq, artifact, imageCfg, downloadBaseURL)
-			}
-		}
-	}
-	if err != nil {
-		failRedoTemplateImageJob(ctx, jobID, resumePhase, err.Error())
-		return
-	}
-
-	readyTargets := targets
-	if resumePhase == JobPhaseDistributing {
-		if err := cleanupArtifactOnNodes(ctx, artifact.ArtifactID, targets); err != nil {
-			failRedoTemplateImageJob(ctx, jobID, JobPhaseDistributing, fmt.Sprintf("cleanup artifact before redistribute failed: %v", err))
-			return
-		}
-		distributedTargets, expected, ready, failed, distErr := distributeRootfsArtifact(ctx, &workingReq, generatedReq, artifact, req.TemplateID, jobID)
-		if err := updateTemplateImageJob(ctx, jobID, map[string]any{
-			"phase":               JobPhaseSnapshotting,
-			"progress":            80,
-			"expected_node_count": expected,
-			"ready_node_count":    ready,
-			"failed_node_count":   failed,
-			"artifact_status":     artifact.Status,
-			"error_message":       errorString(distErr),
-		}); err != nil {
-			logger.Errorf("update redo distribution status fail: %v", err)
-		}
-		if expected > 0 && ready == 0 {
-			_ = updateTemplateImageJob(ctx, jobID, map[string]any{
-				"status":        JobStatusFailed,
-				"phase":         JobPhaseDistributing,
-				"progress":      100,
-				"error_message": fmt.Sprintf("artifact redistribution failed on all %d nodes: %v", expected, distErr),
-			})
-			return
-		}
-		readyTargets = distributedTargets
-		resumePhase = JobPhaseSnapshotting
-	}
-
-	if err := cleanupTemplateReplicasOnNodes(ctx, req.TemplateID, existingReplicas, readyTargets); err != nil {
-		failRedoTemplateImageJob(ctx, jobID, JobPhaseSnapshotting, fmt.Sprintf("cleanup template replicas before redo snapshot failed: %v", err))
-		return
-	}
-	storedReq, err := normalizeStoredTemplateRequest(generatedReq)
-	if err != nil {
-		failRedoTemplateImageJob(ctx, jobID, resumePhase, err.Error())
-		return
-	}
-	if _, err := ensureTemplateDefinition(ctx, req.TemplateID, storedReq, generatedReq.InstanceType, constants.GetAppSnapshotVersion(generatedReq.Annotations)); err != nil {
-		failRedoTemplateImageJob(ctx, jobID, resumePhase, err.Error())
-		return
-	}
-	if _, err := createTemplateReplicasOnNodes(ctx, req.TemplateID, generatedReq, readyTargets, replicaRunOptions{
-		ArtifactID: artifact.ArtifactID,
-		JobID:      jobID,
-	}); err != nil {
-		_ = updateTemplateImageJob(ctx, jobID, map[string]any{
-			"status":          JobStatusFailed,
-			"phase":           JobPhaseSnapshotting,
-			"progress":        100,
-			"template_status": StatusFailed,
-			"error_message":   err.Error(),
-		})
-		return
-	}
-	if err := refreshTemplateReplicaSummary(ctx, req.TemplateID); err != nil {
-		_ = updateTemplateImageJob(ctx, jobID, map[string]any{
-			"status":          JobStatusFailed,
-			"phase":           JobPhaseSnapshotting,
-			"progress":        100,
-			"template_status": StatusFailed,
-			"error_message":   err.Error(),
-		})
-		return
-	}
-	info, err := GetTemplateInfo(ctx, req.TemplateID)
-	if err != nil {
-		_ = updateTemplateImageJob(ctx, jobID, map[string]any{
-			"status":          JobStatusFailed,
-			"phase":           JobPhaseSnapshotting,
-			"progress":        100,
-			"template_status": StatusFailed,
-			"error_message":   err.Error(),
-		})
-		return
-	}
-	resultPayload, _ := json.Marshal(info)
-	finalStatus := JobStatusReady
-	finalPhase := JobPhaseReady
-	if info.Status == StatusFailed {
-		finalStatus = JobStatusFailed
-		finalPhase = JobPhaseSnapshotting
-	}
-	_ = updateTemplateImageJob(ctx, jobID, map[string]any{
-		"status":          finalStatus,
-		"phase":           finalPhase,
-		"progress":        100,
-		"artifact_id":     artifact.ArtifactID,
-		"artifact_status": artifact.Status,
-		"template_status": info.Status,
-		"result_json":     string(resultPayload),
-		"error_message":   info.LastError,
-	})
-}
-
-func ensureRootfsArtifact(ctx context.Context, req *types.CreateTemplateFromImageReq, source *resolvedSourceImage, downloadBaseURL string) (*models.RootfsArtifact, *types.CreateCubeSandboxReq, bool, error) {
-	var generatedReq *types.CreateCubeSandboxReq
-	fingerprint := buildTemplateSpecFingerprint(req, source.digest)
-	artifactID := buildArtifactID(fingerprint)
-	record, wasDeleted, err := findReusableRootfsArtifact(ctx, fingerprint, artifactID)
-	if err == nil && wasDeleted {
-		if restoreErr := restoreRootfsArtifact(ctx, artifactID); restoreErr != nil {
-			return nil, nil, false, restoreErr
-		}
-		record.DeletedAt = gorm.DeletedAt{}
-	}
-	if err == nil && record.Status == ArtifactStatusReady && record.GeneratedRequestJSON != "" {
-		generatedReq, err = generateTemplateCreateRequest(req, record, source.config, downloadBaseURL)
-		if err == nil {
-			return record, generatedReq, false, nil
-		}
-	}
-	if err != nil && !errors.Is(err, gorm.ErrRecordNotFound) {
-		return nil, nil, false, err
-	}
-	if record == nil {
-		record = &models.RootfsArtifact{
-			ArtifactID:              artifactID,
-			TemplateSpecFingerprint: fingerprint,
-			SourceImageRef:          req.SourceImageRef,
-			SourceImageDigest:       source.digest,
-			WritableLayerSize:       req.WritableLayerSize,
-			Status:                  ArtifactStatusPending,
-		}
-		if createErr := store.db.WithContext(ctx).Table(constants.RootfsArtifactTableName).Create(record).Error; createErr != nil {
-			if !errors.Is(createErr, gorm.ErrDuplicatedKey) &&
-				!strings.Contains(createErr.Error(), "1062") &&
-				!strings.Contains(createErr.Error(), "Duplicate entry") {
-				return nil, nil, false, createErr
-			}
-			record, wasDeleted, err = findReusableRootfsArtifact(ctx, fingerprint, artifactID)
-			if err != nil {
-				return nil, nil, false, createErr
-			}
-			if wasDeleted {
-				if restoreErr := restoreRootfsArtifact(ctx, artifactID); restoreErr != nil {
-					return nil, nil, false, restoreErr
-				}
-				record.DeletedAt = gorm.DeletedAt{}
-			}
-			if record.Status == ArtifactStatusReady && record.GeneratedRequestJSON != "" {
-				generatedReq, err = generateTemplateCreateRequest(req, record, source.config, downloadBaseURL)
-				if err == nil {
-					return record, generatedReq, false, nil
-				}
-			}
-		}
-	}
-	_ = updateRootfsArtifact(ctx, artifactID, map[string]any{
-		"template_spec_fingerprint": fingerprint,
-		"source_image_ref":          req.SourceImageRef,
-		"source_image_digest":       source.digest,
-		"writable_layer_size":       req.WritableLayerSize,
-		"status":                    ArtifactStatusBuilding,
-		"last_error":                "",
-	})
-	record, generatedReq, err = buildRootfsArtifact(ctx, record, req, source, downloadBaseURL)
-	if err != nil {
-		_ = updateRootfsArtifact(ctx, artifactID, map[string]any{
-			"status":     ArtifactStatusFailed,
-			"last_error": err.Error(),
-		})
-		return nil, nil, false, err
-	}
-	return record, generatedReq, true, nil
-}
-
-func findReusableRootfsArtifact(ctx context.Context, fingerprint, artifactID string) (*models.RootfsArtifact, bool, error) {
-	record, err := getRootfsArtifactByFingerprint(ctx, fingerprint)
-	if err == nil {
-		record, err = validateReusableRootfsArtifact(record, fingerprint, artifactID)
-		return record, rootfsArtifactSoftDeleted(record), err
-	}
-	if !errors.Is(err, gorm.ErrRecordNotFound) {
-		return nil, false, err
-	}
-
-	record, err = getRootfsArtifactByFingerprintUnscoped(ctx, fingerprint)
-	if err == nil {
-		record, err = validateReusableRootfsArtifact(record, fingerprint, artifactID)
-		return record, rootfsArtifactSoftDeleted(record), err
-	}
-	if !errors.Is(err, gorm.ErrRecordNotFound) {
-		return nil, false, err
-	}
-
-	record, err = getRootfsArtifactByID(ctx, artifactID)
-	if err != nil {
-		record, err = getRootfsArtifactByIDUnscoped(ctx, artifactID)
-		if err != nil {
-			return nil, false, err
-		}
-	}
-	record, err = validateReusableRootfsArtifact(record, fingerprint, artifactID)
-	return record, rootfsArtifactSoftDeleted(record), err
-}
-
-func validateReusableRootfsArtifact(record *models.RootfsArtifact, fingerprint, artifactID string) (*models.RootfsArtifact, error) {
-	if record == nil {
-		return nil, gorm.ErrRecordNotFound
-	}
-	if record.ArtifactID != artifactID {
-		return nil, fmt.Errorf("rootfs artifact id mismatch: want %s got %s", artifactID, record.ArtifactID)
-	}
-	if record.TemplateSpecFingerprint != "" && record.TemplateSpecFingerprint != fingerprint {
-		return nil, fmt.Errorf("rootfs artifact %s fingerprint mismatch: want %s got %s", artifactID, fingerprint, record.TemplateSpecFingerprint)
-	}
-	return record, nil
-}
-
-func rootfsArtifactSoftDeleted(record *models.RootfsArtifact) bool {
-	return record != nil && record.DeletedAt.Valid
-}
-
-func restoreRootfsArtifact(ctx context.Context, artifactID string) error {
-	tx := store.db.WithContext(ctx).Unscoped().Table(constants.RootfsArtifactTableName).
-		Where("artifact_id = ?", artifactID).
-		Updates(map[string]any{
-			"deleted_at": nil,
-			"updated_at": time.Now(),
-		})
-	if tx.Error != nil {
-		return tx.Error
-	}
-	if tx.RowsAffected == 0 {
-		return gorm.ErrRecordNotFound
-	}
-	return nil
-}
-
-func buildRootfsArtifact(ctx context.Context, record *models.RootfsArtifact, req *types.CreateTemplateFromImageReq, source *resolvedSourceImage, downloadBaseURL string) (*models.RootfsArtifact, *types.CreateCubeSandboxReq, error) {
-	workDir := filepath.Join(artifactWorkRootDir(), record.ArtifactID)
-	storeDir, err := resolveArtifactStoreDir(ctx, record.ArtifactID)
-	if err != nil {
-		return nil, nil, err
-	}
-	storeRootfsDir := filepath.Join(storeDir, "rootfs")
-	ext4Path := filepath.Join(storeDir, record.ArtifactID+".ext4")
-	keepStoreDir := false
-
-	// Phase 2: loop-mount streaming build (optional, auto-detects capability).
-	//
-	// The streaming build relies on `docker create` + `docker export`, so it is
-	// only available on the docker path. For dockerless sources (skopeo+umoci)
-	// we deliberately skip Phase 2 and fall through to the daemonless Phase 1 to
-	// keep the dockerless path free of any docker binary dependency. umoci
-	// unpacks to an on-disk bundle rather than streaming to stdout, so a
-	// loop-mount "streaming" variant would not avoid the intermediate rootfs
-	// anyway.
-	if loopMountExt4Enabled() && canUseLoopMount() && !source.useDockerless {
-		estimatedPhase2, err := estimateImageSizeFromInspect(ctx, source)
-		if err != nil {
-			log.G(ctx).Warnf("cannot estimate image size for Phase 2, falling back to Phase 1: %v", err)
-		} else {
-			if err := checkDiskSpace(ctx, storeDir, estimatedPhase2); err != nil {
-				return nil, nil, err
-			}
-			if err := createExt4ImageStreaming(ctx, source, workDir, ext4Path, estimatedPhase2); err != nil {
-				log.G(ctx).Warnf("loop-mount streaming ext4 build failed, falling back to phase-1: %v", err)
-				// Clean up partial work and fall through to phase-1 path.
-				_ = os.RemoveAll(workDir)
-				_ = os.Remove(ext4Path)
-			} else {
-				// Streaming build succeeded – compute SHA256, record metadata, and return.
-				shaValue, sizeBytes, err := computeFileSHA256(ext4Path)
-				if err != nil {
-					return nil, nil, err
-				}
-				_ = os.RemoveAll(workDir) // clean up temp work directory on success
-				keepStoreDir = true
-				return finalizeArtifact(ctx, record, source, ext4Path, shaValue, sizeBytes, downloadBaseURL, req)
-			}
-		}
-	}
-
-	// Phase 1: disk space pre-check.
-	// Size estimate comes from skopeo inspect (dockerless) or docker image
-	// inspect (docker), selected by estimateImageSizeFromInspect.
-	estimatedSizeBytes, err := estimateImageSizeFromInspect(ctx, source)
-	if err != nil {
-		log.G(ctx).Warnf("cannot estimate image size for disk-space check, skipping: %v", err)
-	} else if estimatedSizeBytes > 0 {
-		if err := checkDiskSpace(ctx, storeDir, estimatedSizeBytes); err != nil {
-			return nil, nil, err
-		}
-	}
-
-	// Defer cleanup of workDir.  storeDir cleanup is handled explicitly below.
-	defer func() {
-		if workDir != "" {
-			if err := os.RemoveAll(workDir); err != nil { // NOCC:Path Traversal()
-				log.G(ctx).Warnf("cleanup workDir %s failed: %v", workDir, err)
-			}
-		}
-		if !keepStoreDir {
-			if storeDir != "" {
-				if err := os.RemoveAll(storeDir); err != nil { // NOCC:Path Traversal()
-					log.G(ctx).Warnf("cleanup storeDir %s failed: %v", storeDir, err)
-				}
-			}
-		} else {
-			if storeRootfsDir != "" {
-				if err := os.RemoveAll(storeRootfsDir); err != nil { // NOCC:Path Traversal()
-					log.G(ctx).Warnf("cleanup storeRootfsDir %s failed: %v", storeRootfsDir, err)
-				}
-			}
-		}
-	}()
-
-	// Optimisation 4: export directly to storeDir when on a local fast filesystem.
-	// When storeDir is on NFS/CIFS, fall back to workDir + relocate.
-	if isLocalFastFS(storeDir) {
-		if err := exportImageRootfs(ctx, source, storeRootfsDir); err != nil {
-			return nil, nil, err
-		}
-	} else {
-		rootfsDir := filepath.Join(workDir, "rootfs")
-		if err := os.MkdirAll(rootfsDir, 0o755); err != nil {
-			return nil, nil, err
-		}
-		if err := exportImageRootfs(ctx, source, rootfsDir); err != nil {
-			return nil, nil, err
-		}
-		if err := relocateRootfsToArtifactStore(ctx, rootfsDir, storeRootfsDir); err != nil {
-			return nil, nil, err
-		}
-	}
-
-	// Optimisation 2: clean up workDir early – at this point rootfs has been
-	// exported to storeRootfsDir and workDir is no longer needed for ext4 creation.
-	if workDir != "" {
-		if err := os.RemoveAll(workDir); err != nil { // NOCC:Path Traversal()
-			log.G(ctx).Warnf("cleanup workDir %s failed: %v", workDir, err)
-		}
-	}
-
-	if err := createExt4Image(ctx, storeRootfsDir, ext4Path); err != nil {
-		return nil, nil, err
-	}
-	shaValue, sizeBytes, err := computeFileSHA256(ext4Path)
-	if err != nil {
-		return nil, nil, err
-	}
-	keepStoreDir = true
-	return finalizeArtifact(ctx, record, source, ext4Path, shaValue, sizeBytes, downloadBaseURL, req)
-}
-
-// finalizeArtifact populates the artifact record with computed values, persists it,
-// and returns the latest version.
-func finalizeArtifact(ctx context.Context, record *models.RootfsArtifact, source *resolvedSourceImage, ext4Path, shaValue string, sizeBytes int64, downloadBaseURL string, req *types.CreateTemplateFromImageReq) (*models.RootfsArtifact, *types.CreateCubeSandboxReq, error) {
-	downloadToken := uuid.New().String()
-	record.SourceImageDigest = source.digest
-	record.MasterNodeIP = source.masterNodeIP
-	record.Ext4Path = ext4Path
-	record.Ext4SHA256 = shaValue
-	record.Ext4SizeBytes = sizeBytes
-	record.ImageConfigJSON = source.configJSON
-	record.DownloadToken = downloadToken
-	record.Status = ArtifactStatusReady
-	record.GCDeadline = time.Now().Add(defaultTemplateArtifactTTL).Unix()
-
-	generatedReq, err := generateTemplateCreateRequest(req, record, source.config, downloadBaseURL)
-	if err != nil {
-		return nil, nil, err
-	}
-	reqPayload, err := json.Marshal(generatedReq)
-	if err != nil {
-		return nil, nil, err
-	}
-	record.GeneratedRequestJSON = string(reqPayload)
-	if err := updateRootfsArtifact(ctx, record.ArtifactID, map[string]any{
-		"source_image_digest":    record.SourceImageDigest,
-		"master_node_ip":         record.MasterNodeIP,
-		"ext4_path":              record.Ext4Path,
-		"ext4_sha256":            record.Ext4SHA256,
-		"ext4_size_bytes":        record.Ext4SizeBytes,
-		"image_config_json":      record.ImageConfigJSON,
-		"generated_request_json": record.GeneratedRequestJSON,
-		"download_token":         record.DownloadToken,
-		"status":                 record.Status,
-		"gc_deadline":            record.GCDeadline,
-		"last_error":             "",
-	}); err != nil {
-		return nil, nil, err
-	}
-	latest, err := getRootfsArtifactByID(ctx, record.ArtifactID)
-	if err != nil {
-		return nil, nil, err
-	}
-	return latest, generatedReq, nil
-}
-
-func prepareSourceImage(ctx context.Context, req *types.CreateTemplateFromImageReq, downloadBaseURL string) (*resolvedSourceImage, error) {
-	if req == nil {
-		return nil, errors.New("request is nil")
-	}
-	if hasDockerlessRootfsExportTools() {
-		return prepareDockerlessSourceImage(ctx, req, downloadBaseURL)
-	}
-	return prepareDockerSourceImage(ctx, req, downloadBaseURL)
-}
-
-func prepareDockerlessSourceImage(ctx context.Context, req *types.CreateTemplateFromImageReq, downloadBaseURL string) (*resolvedSourceImage, error) {
-	if req == nil {
-		return nil, errors.New("request is nil")
-	}
-	if err := validateImageRef(req.SourceImageRef); err != nil {
-		return nil, err
-	}
-	authFile, cleanup, err := createSkopeoAuthFile(req.SourceImageRef, req.RegistryUsername, req.RegistryPassword)
-	if err != nil {
-		return nil, err
-	}
-	sourceRef := skopeoDockerImageRef(req.SourceImageRef)
-	inspectOutput, err := skopeoOutput(ctx, authFile, "inspect", sourceRef)
-	if err != nil {
-		cleanup()
-		return nil, fmt.Errorf("skopeo inspect %s failed: %w", req.SourceImageRef, err)
-	}
-	configOutput, err := skopeoOutput(ctx, authFile, "inspect", "--config", sourceRef)
-	if err != nil {
-		cleanup()
-		return nil, fmt.Errorf("skopeo inspect --config %s failed: %w", req.SourceImageRef, err)
-	}
-
-	var inspectInfo skopeoInspectImage
-	if err := json.Unmarshal(inspectOutput, &inspectInfo); err != nil {
-		cleanup()
-		return nil, fmt.Errorf("unmarshal skopeo inspect output: %w", err)
-	}
-	var configInfo skopeoInspectConfig
-	if err := json.Unmarshal(configOutput, &configInfo); err != nil {
-		cleanup()
-		return nil, fmt.Errorf("unmarshal skopeo inspect config output: %w", err)
-	}
-	configJSON, err := json.Marshal(configInfo.Config)
-	if err != nil {
-		cleanup()
-		return nil, fmt.Errorf("marshal image config: %w", err)
-	}
-	return &resolvedSourceImage{
-		localRef:            req.SourceImageRef,
-		digest:              skopeoImageDigest(inspectInfo, req.SourceImageRef),
-		config:              configInfo.Config,
-		configJSON:          string(configJSON),
-		masterNodeIP:        normalizeBaseURL(downloadBaseURL),
-		useDockerless:       true,
-		skopeoAuthFile:      authFile,
-		compressedSizeBytes: skopeoLayersTotalSize(inspectInfo),
-		cleanup: func(context.Context) {
-			cleanup()
-		},
-	}, nil
-}
-
-func prepareDockerSourceImage(ctx context.Context, req *types.CreateTemplateFromImageReq, downloadBaseURL string) (*resolvedSourceImage, error) {
-	if req == nil {
-		return nil, errors.New("request is nil")
-	}
-	var (
-		dockerConfigDir       string
-		removeDockerConfigDir bool
-		imageExistsLocally    bool
-		inspectOutput         []byte
-		err                   error
-	)
-	defer func() {
-		if removeDockerConfigDir && dockerConfigDir != "" {
-			_ = os.RemoveAll(dockerConfigDir)
-		}
-	}()
-	inspectOutput, err = dockerOutput(ctx, "", "image", "inspect", "--", req.SourceImageRef)
-	if err == nil {
-		imageExistsLocally = true
-	}
-	if req.RegistryUsername != "" || req.RegistryPassword != "" {
-		tmpDir, err := os.MkdirTemp("", "cubemaster-docker-config-*")
-		if err != nil {
-			return nil, err
-		}
-		dockerConfigDir = tmpDir
-		removeDockerConfigDir = true
-		if err := dockerLogin(ctx, dockerConfigDir, req.SourceImageRef, req.RegistryUsername, req.RegistryPassword); err != nil {
-			return nil, err
-		}
-	}
-	if !imageExistsLocally {
-		if err := dockerRun(ctx, dockerConfigDir, "pull", "--", req.SourceImageRef); err != nil {
-			return nil, fmt.Errorf("docker pull %s failed: %w", req.SourceImageRef, err)
-		}
-		inspectOutput, err = dockerOutput(ctx, dockerConfigDir, "image", "inspect", "--", req.SourceImageRef)
-		if err != nil {
-			return nil, fmt.Errorf("docker image inspect %s failed: %w", req.SourceImageRef, err)
-		}
-	}
-	var inspectList []dockerInspectImage
-	if err := json.Unmarshal(inspectOutput, &inspectList); err != nil {
-		return nil, fmt.Errorf("unmarshal docker inspect output: %w", err)
-	}
-	if len(inspectList) == 0 {
-		return nil, fmt.Errorf("docker image inspect returned empty result for %s", req.SourceImageRef)
-	}
-	inspectInfo := inspectList[0]
-	configJSON, err := json.Marshal(inspectInfo.Config)
-	if err != nil {
-		return nil, fmt.Errorf("marshal image config: %w", err)
-	}
-	source := &resolvedSourceImage{
-		localRef:     req.SourceImageRef,
-		digest:       firstNonEmptyDigest(inspectInfo),
-		config:       inspectInfo.Config,
-		configJSON:   string(configJSON),
-		masterNodeIP: normalizeBaseURL(downloadBaseURL),
-		cleanup: func(cleanupCtx context.Context) {
-			if dockerConfigDir != "" {
-				_ = os.RemoveAll(dockerConfigDir)
-			}
-			if !imageExistsLocally {
-				_ = dockerRun(cleanupCtx, "", "image", "rm", "-f", "--", req.SourceImageRef)
-			}
-		},
-	}
-	removeDockerConfigDir = false
-	return source, nil
-}
-
-// imageRefAllowedPattern is the strict character whitelist for image
-// references. It permits exactly the characters that appear in legitimate
-// registry/repository[:tag][@algo:hexdigest] references: alphanumerics and
-// `.`, `-`, `_`, `/`, `:`, `@`. Any other character (notably whitespace) is
-// rejected so the reference cannot be split into additional argv entries.
-var imageRefAllowedPattern = regexp.MustCompile(`^[A-Za-z0-9._:/@-]+$`)
-
-// validateImageRef guards against argument injection (CWE-88) when the image
-// reference is later passed as a positional argument to external CLIs
-// (skopeo/umoci/docker). Those tools accept flags interspersed with positional
-// arguments, so a ref such as `registry.example.com/image --authfile /etc/shadow`
-// would otherwise smuggle extra flags into the subprocess. To prevent this we
-// enforce a strict character whitelist (which excludes whitespace and other
-// argument delimiters) and reject any ref that begins with a dash.
-func validateImageRef(imageRef string) error {
-	trimmed := strings.TrimPrefix(imageRef, "docker://")
-	if trimmed == "" {
-		return errors.New("empty image reference")
-	}
-	if strings.HasPrefix(trimmed, "-") {
-		return fmt.Errorf("invalid image reference: %s", imageRef)
-	}
-	if !imageRefAllowedPattern.MatchString(trimmed) {
-		return fmt.Errorf("invalid image reference: %s", imageRef)
-	}
-	return nil
-}
-
-func exportImageRootfs(ctx context.Context, source *resolvedSourceImage, destRootfsDir string) error {
-	if source == nil {
-		return errors.New("resolved source image is nil")
-	}
-	if err := validateImageRef(source.localRef); err != nil {
-		return err
-	}
-	// Use the strategy chosen at prepare time rather than re-detecting, so the
-	// prepare and export phases never diverge (see resolvedSourceImage.useDockerless).
-	if source.useDockerless {
-		return dockerlessExportImageRootfs(ctx, source, destRootfsDir)
-	}
-	return dockerExportImageRootfs(ctx, source, destRootfsDir)
-}
-
-// dockerlessExportImageRootfs exports the source image's root filesystem
-// without a docker daemon. The flow is:
-//
-//  1. `skopeo copy docker://<ref> oci:<workDir>/image` pulls the image into a
-//     local OCI layout (honoring the optional skopeo auth file).
-//  2. `umoci unpack --rootless` materializes that layout into an OCI bundle
-//     under <workDir>/bundle, whose `rootfs` subdir holds the extracted files.
-//  3. The bundle's rootfs is moved into place at destRootfsDir via os.Rename.
-//
-// The scratch workDir is created under destRootfsDir's parent (rather than
-// $TMPDIR) so the final os.Rename stays on the same filesystem and cannot fail
-// with EXDEV, which would otherwise force a slow full copy across devices.
-func dockerlessExportImageRootfs(ctx context.Context, source *resolvedSourceImage, destRootfsDir string) error {
-	if err := os.MkdirAll(filepath.Dir(destRootfsDir), 0o755); err != nil {
-		return err
-	}
-	workDir, err := os.MkdirTemp(filepath.Dir(destRootfsDir), ".dockerless-rootfs-*")
-	if err != nil {
-		return err
-	}
-	defer os.RemoveAll(workDir)
-
-	ociDir := filepath.Join(workDir, "image")
-	bundleDir := filepath.Join(workDir, "bundle")
-
-	sourceRef := skopeoDockerImageRef(source.localRef)
-	ociImageRef := "oci:" + ociLayoutImageRef(ociDir, source.localRef)
-	skopeoArgs := []string{"copy"}
-	if source.skopeoAuthFile != "" {
-		skopeoArgs = append(skopeoArgs, "--authfile", source.skopeoAuthFile)
-	}
-	skopeoArgs = append(skopeoArgs, sourceRef, ociImageRef)
-	if err := runCommand(ctx, "", "skopeo", skopeoArgs...); err != nil {
-		return fmt.Errorf("skopeo copy %s failed: %w", source.localRef, err)
-	}
-
-	umociImageRef := ociLayoutImageRef(ociDir, source.localRef)
-	if err := runCommand(ctx, "", "umoci", "unpack", "--rootless", "--image", umociImageRef, bundleDir); err != nil {
-		return fmt.Errorf("umoci unpack %s failed: %w", source.localRef, err)
-	}
-	// The OCI layout is no longer needed once unpacked; drop it early to reduce
-	// peak disk usage on the artifact filesystem.
-	_ = os.RemoveAll(ociDir) // NOCC:Path Traversal()
-
-	unpackedRootfsDir := filepath.Join(bundleDir, "rootfs")
-	if err := os.RemoveAll(destRootfsDir); err != nil { // NOCC:Path Traversal()
-		return err
-	}
-	if err := os.Rename(unpackedRootfsDir, destRootfsDir); err != nil { // NOCC:Path Traversal()
-		return fmt.Errorf("move unpacked rootfs failed: %w", err)
-	}
-	return nil
-}
-
-// hasDockerlessRootfsExportTools reports whether both skopeo and umoci are
-// available on PATH. When they are, the build prefers the daemonless export
-// path (skopeo+umoci) over docker. This auto-detection means installing both
-// tools on a cubemaster node silently switches the source-image handling away
-// from the docker daemon; ensureArtifactBuildPreflight reflects the resulting
-// command requirements.
-func hasDockerlessRootfsExportTools() bool {
-	if _, err := executableLookPath("skopeo"); err != nil {
-		return false
-	}
-	if _, err := executableLookPath("umoci"); err != nil {
-		return false
-	}
-	return true
-}
-
-func skopeoDockerImageRef(imageRef string) string {
-	if strings.HasPrefix(imageRef, "docker://") {
-		return imageRef
-	}
-	return "docker://" + imageRef
-}
-
-func createSkopeoAuthFile(imageRef, username, password string) (string, func(), error) {
-	if username == "" && password == "" {
-		return "", func() {}, nil
-	}
-	tmpDir, err := os.MkdirTemp("", "cubemaster-skopeo-auth-*")
-	if err != nil {
-		return "", nil, err
-	}
-	cleanup := func() {
-		_ = os.RemoveAll(tmpDir)
-	}
-	authPayload := map[string]any{
-		"auths": map[string]any{
-			registryHostFromImageRef(imageRef): map[string]string{
-				"auth": base64.StdEncoding.EncodeToString([]byte(username + ":" + password)),
-			},
-		},
-	}
-	payload, err := json.Marshal(authPayload)
-	if err != nil {
-		cleanup()
-		return "", nil, err
-	}
-	authFile := filepath.Join(tmpDir, "auth.json")
-	if err := os.WriteFile(authFile, payload, 0o600); err != nil {
-		cleanup()
-		return "", nil, err
-	}
-	return authFile, cleanup, nil
-}
-
-func skopeoImageDigest(info skopeoInspectImage, imageRef string) string {
-	if info.Digest == "" {
-		return ""
-	}
-	name := info.Name
-	if name == "" {
-		name = imageNameWithoutTagDigest(imageRef)
-	}
-	if name == "" {
-		return info.Digest
-	}
-	return name + "@" + info.Digest
-}
-
-func ociLayoutImageRef(ociDir, imageRef string) string {
-	tag := imageTagFromRef(imageRef)
-	if tag == "" {
-		return ociDir
-	}
-	return ociDir + ":" + tag
-}
-
-// splitImageRef strips the docker:// transport prefix and any @digest suffix,
-// then splits the remainder into the repository name and the tag. The final
-// ":" is only treated as a tag separator when it appears after the last "/", so
-// a registry port (e.g. registry:5000/image) is not mistaken for a tag. When
-// the reference carries no tag, name is the whole remainder and tag is empty.
-func splitImageRef(imageRef string) (name, tag string) {
-	imageRef = strings.TrimPrefix(imageRef, "docker://")
-	if digestIndex := strings.LastIndex(imageRef, "@"); digestIndex >= 0 {
-		imageRef = imageRef[:digestIndex]
-	}
-	lastSlash := strings.LastIndex(imageRef, "/")
-	lastColon := strings.LastIndex(imageRef, ":")
-	if lastColon > lastSlash {
-		return imageRef[:lastColon], imageRef[lastColon+1:]
-	}
-	return imageRef, ""
-}
-
-func imageTagFromRef(imageRef string) string {
-	_, tag := splitImageRef(imageRef)
-	return tag
-}
-
-func imageNameWithoutTagDigest(imageRef string) string {
-	name, _ := splitImageRef(imageRef)
-	return name
-}
-
-func dockerExportImageRootfs(ctx context.Context, source *resolvedSourceImage, destRootfsDir string) error {
-	containerIDBytes, err := dockerOutput(ctx, "", "create", "--", source.localRef)
-	if err != nil {
-		return fmt.Errorf("docker create %s failed: %w", source.localRef, err)
-	}
-	containerID := strings.TrimSpace(string(containerIDBytes))
-	// Use context.Background() so that cleanup runs even after request ctx is cancelled.
-	cleanupCtx := context.Background()
-	defer func() {
-		_ = dockerRun(cleanupCtx, "", "rm", "-f", containerID)
-	}()
-
-	if err := os.RemoveAll(destRootfsDir); err != nil { // NOCC:Path Traversal()
-		return err
-	}
-	if err := os.MkdirAll(destRootfsDir, 0o755); err != nil {
-		return err
-	}
-
-	return pipeExportToDir(ctx, containerID, destRootfsDir)
-}
-
-func createExt4Image(ctx context.Context, rootfsDir, ext4Path string) error {
-	sizeBytes, fileCount, err := directorySizeAndFileCount(rootfsDir)
-	if err != nil {
-		return err
-	}
-
-	const mib = int64(1024 * 1024)
-	const gib = int64(1024 * 1024 * 1024)
-
-	// Fixed overhead (default 256 MiB, configurable).
-	fixedOverhead := ext4FixedOverheadMiB() * mib
-
-	// Percentage overhead: configurable percentage of the data size (default 10%).
-	percentageOverhead := sizeBytes * ext4OverheadPercent() / 100
-
-	// Per-file overhead: ~1 KiB per file for inode (256 B) + directory entry + indirect block alignment.
-	perFileOverhead := fileCount * 1024
-
-	raw := sizeBytes + fixedOverhead + percentageOverhead + perFileOverhead
-
-	// Minimum 1 GiB.
-	if raw < gib {
-		raw = gib
-	}
-
-	// Align up to 256 MiB boundary instead of next power-of-2.
-	alignment := int64(256) * mib
-	imageSize := ((raw + alignment - 1) / alignment) * alignment
-
-	if err := runCommand(ctx, "", "truncate", "-s", strconv.FormatInt(imageSize, 10), ext4Path); err != nil {
-		return fmt.Errorf("truncate ext4 image failed: %w", err)
-	}
-	if err := runCommand(ctx, "", "mkfs.ext4", "-F", "-d", rootfsDir, ext4Path); err != nil {
-		return fmt.Errorf("mkfs.ext4 failed: %w", err)
-	}
-	return nil
-}
-
-func generateTemplateCreateRequest(req *types.CreateTemplateFromImageReq, artifact *models.RootfsArtifact, imageCfg dockerImageConfig, downloadBaseURL string) (*types.CreateCubeSandboxReq, error) {
-	annotations := map[string]string{
-		constants.CubeAnnotationAppSnapshotTemplateID:      req.TemplateID,
-		constants.CubeAnnotationsAppSnapshotCreate:         "true",
-		constants.CubeAnnotationAppSnapshotVersion:         DefaultTemplateVersion,
-		constants.CubeAnnotationAppSnapshotTemplateVersion: DefaultTemplateVersion,
-		constants.CubeAnnotationRootfsArtifactID:           artifact.ArtifactID,
-		constants.CubeAnnotationWritableLayerSize:          req.WritableLayerSize,
-		constants.CubeAnnotationTemplateSpecFingerprint:    artifact.TemplateSpecFingerprint,
-	}
-	sizeGi, err := quantityToGi(req.WritableLayerSize)
-	if err == nil && sizeGi > 0 {
-		annotations[constants.CubeAnnotationsSystemDiskSize] = strconv.FormatInt(sizeGi, 10)
-	}
-	if len(req.ExposedPorts) > 0 {
-		annotations[constants.AnnotationsExposedPort] = formatExposedPortsAnnotation(req.ExposedPorts)
-	}
-	rootVolume := &types.Volume{
-		Name: rootfsWritableVolumeName,
-		VolumeSource: &types.VolumeSource{
-			EmptyDir: &types.EmptyDirVolumeSource{
-				SizeLimit: req.WritableLayerSize,
-			},
-		},
-	}
-	imageAnnotations := map[string]string{
-		constants.CubeAnnotationRootfsArtifactID:        artifact.ArtifactID,
-		constants.CubeAnnotationRootfsArtifactURL:       buildDownloadURL(downloadBaseURL, artifact.ArtifactID, artifact.DownloadToken),
-		constants.CubeAnnotationRootfsArtifactToken:     artifact.DownloadToken,
-		constants.CubeAnnotationRootfsArtifactSHA256:    artifact.Ext4SHA256,
-		constants.CubeAnnotationRootfsArtifactSizeBytes: strconv.FormatInt(artifact.Ext4SizeBytes, 10),
-		constants.CubeAnnotationWritableLayerSize:       req.WritableLayerSize,
-		constants.CubeAnnotationTemplateSpecFingerprint: artifact.TemplateSpecFingerprint,
-	}
-	command := imageCfg.Entrypoint
-	args := imageCfg.Cmd
-	if req.ContainerOverrides != nil {
-		if len(req.ContainerOverrides.Command) > 0 {
-			command = req.ContainerOverrides.Command
-		}
-		if len(req.ContainerOverrides.Args) > 0 {
-			args = req.ContainerOverrides.Args
-		}
-	}
-	envs := envListToKeyValues(imageCfg.Env)
-	if req.ContainerOverrides != nil && req.ContainerOverrides.Envs != nil {
-		envs = req.ContainerOverrides.Envs
-	}
-	workingDir := imageCfg.WorkingDir
-	if req.ContainerOverrides != nil && req.ContainerOverrides.WorkingDir != "" {
-		workingDir = req.ContainerOverrides.WorkingDir
-	}
-	resources := &types.Resource{Cpu: defaultTemplateCPU, Mem: defaultTemplateMemory}
-	if req.ContainerOverrides != nil && req.ContainerOverrides.Resources != nil {
-		resources = req.ContainerOverrides.Resources
-	}
-	securityContext := &types.ContainerSecurityContext{Privileged: true, ReadonlyRootfs: false}
-	if req.ContainerOverrides != nil && req.ContainerOverrides.SecurityContext != nil {
-		securityContext = req.ContainerOverrides.SecurityContext
-		securityContext.ReadonlyRootfs = false
-	}
-	if req.ContainerOverrides != nil && req.ContainerOverrides.VolumeMounts != nil {
-		for _, mount := range req.ContainerOverrides.VolumeMounts {
-			if mount != nil && mount.ContainerPath == "/" {
-				return nil, fmt.Errorf("container_overrides.volume_mounts must not override / because writable rootfs is template-owned")
-			}
-		}
-	}
-	volumeMounts := []*cubeboxv1.VolumeMounts{{
-		Name:          rootfsWritableVolumeName,
-		ContainerPath: "/",
-	}}
-	if req.ContainerOverrides != nil && len(req.ContainerOverrides.VolumeMounts) > 0 {
-		volumeMounts = append(volumeMounts, req.ContainerOverrides.VolumeMounts...)
-	}
-	containerAnnotations := map[string]string{}
-	if req.ContainerOverrides != nil && req.ContainerOverrides.Annotations != nil {
-		for k, v := range req.ContainerOverrides.Annotations {
-			containerAnnotations[k] = v
-		}
-	}
-	container := &types.Container{
-		Name:            "cubebox-name-0",
-		Image:           &types.ImageSpec{Image: artifact.ArtifactID, StorageMedia: imagev1.ImageStorageMediaType_ext4.String(), WritableLayerSize: req.WritableLayerSize, Annotations: imageAnnotations},
-		Command:         command,
-		Args:            args,
-		WorkingDir:      workingDir,
-		Envs:            envs,
-		VolumeMounts:    volumeMounts,
-		DnsConfig:       dnsConfigOrNil(req.ContainerOverrides),
-		RLimit:          defaultRLimit(req.ContainerOverrides),
-		Resources:       resources,
-		SecurityContext: securityContext,
-		Probe:           probeOrNil(req.ContainerOverrides),
-		Annotations:     containerAnnotations,
-	}
-	return &types.CreateCubeSandboxReq{
-		Request:       &types.Request{RequestID: req.RequestID},
-		Volumes:       []*types.Volume{rootVolume},
-		Containers:    []*types.Container{container},
-		Annotations:   annotations,
-		InstanceType:  req.InstanceType,
-		NetworkType:   req.NetworkType,
-		CubeVSContext: cloneCubeVSContext(req.CubeVSContext),
-	}, nil
-}
-
-func cloneCubeVSContext(in *types.CubeVSContext) *types.CubeVSContext {
-	if in == nil {
-		return nil
-	}
-	out := &types.CubeVSContext{
-		AllowOut: append([]string(nil), in.AllowOut...),
-		DenyOut:  append([]string(nil), in.DenyOut...),
-	}
-	if in.AllowInternetAccess != nil {
-		allowInternetAccess := *in.AllowInternetAccess
-		out.AllowInternetAccess = &allowInternetAccess
-	}
-	return out
-}
-
-func formatTemplateImageCubeVSContext(in *types.CubeVSContext) string {
-	if in == nil {
-		return "allow_internet_access=default(true) allow_out=[] deny_out=[]"
-	}
-	allowInternetAccess := "default(true)"
-	if in.AllowInternetAccess != nil {
-		allowInternetAccess = fmt.Sprintf("%t", *in.AllowInternetAccess)
-	}
-	return fmt.Sprintf("allow_internet_access=%s allow_out=%v deny_out=%v", allowInternetAccess, in.AllowOut, in.DenyOut)
-}
-
-func distributeRootfsArtifact(ctx context.Context, req *types.CreateTemplateFromImageReq, generatedReq *types.CreateCubeSandboxReq, artifact *models.RootfsArtifact, templateID, jobID string) ([]*node.Node, int32, int32, int32, error) {
-	targets, err := resolveTemplateNodes(req.InstanceType, req.DistributionScope)
-	if err != nil {
-		return nil, 0, 0, 0, err
-	}
-	spec := &imagev1.ImageSpec{
-		Image:        artifact.ArtifactID,
-		StorageMedia: imagev1.ImageStorageMediaType_ext4.String(),
-		Annotations: map[string]string{
-			constants.CubeAnnotationRootfsArtifactID:        artifact.ArtifactID,
-			constants.CubeAnnotationRootfsArtifactURL:       buildDownloadURL(artifact.MasterNodeIP, artifact.ArtifactID, artifact.DownloadToken),
-			constants.CubeAnnotationRootfsArtifactToken:     artifact.DownloadToken,
-			constants.CubeAnnotationRootfsArtifactSHA256:    artifact.Ext4SHA256,
-			constants.CubeAnnotationRootfsArtifactSizeBytes: strconv.FormatInt(artifact.Ext4SizeBytes, 10),
-			constants.CubeAnnotationWritableLayerSize:       req.WritableLayerSize,
-			constants.CubeAnnotationTemplateSpecFingerprint: artifact.TemplateSpecFingerprint,
-			constants.CubeAnnotationsInsType:                req.InstanceType,
-		},
-	}
-	expected := int32(len(targets))
-	ready := int32(0)
-	failed := int32(0)
-	var firstErr error
-	var lock sync.Mutex
-	sem := make(chan struct{}, defaultDistributionWorkers)
-	var wg sync.WaitGroup
-	readyTargets := make([]*node.Node, 0, len(targets))
-	for _, target := range targets {
-		target := target
-		wg.Add(1)
-		go func() {
-			defer wg.Done()
-			sem <- struct{}{}
-			defer func() { <-sem }()
-
-			replica := buildReplicaForDistribution(target, generatedReq, artifact.ArtifactID, jobID)
-			rsp, err := cubelet.CreateImage(ctx, cubelet.GetCubeletAddr(target.HostIP()), &imagev1.CreateImageRequest{
-				RequestID: uuid.New().String(),
-				Spec:      spec,
-			})
-			lock.Lock()
-			defer lock.Unlock()
-			if err != nil {
-				failed++
-				replica.Phase = ReplicaPhaseFailed
-				replica.ErrorMessage = err.Error()
-				if firstErr == nil {
-					firstErr = err
-				}
-				if templateID != "" && generatedReq != nil {
-					_ = UpsertReplica(ctx, templateID, generatedReq.InstanceType, replica)
-				}
-				return
-			}
-			if rsp.GetRet() == nil || int(rsp.GetRet().GetRetCode()) != int(errorcode.ErrorCode_Success) {
-				failed++
-				replica.Phase = ReplicaPhaseFailed
-				if firstErr == nil {
-					if rsp.GetRet() != nil {
-						firstErr = fmt.Errorf("cubelet create image on %s failed: %s", target.HostIP(), rsp.GetRet().GetRetMsg())
-					} else {
-						firstErr = fmt.Errorf("cubelet create image on %s returned empty ret", target.HostIP())
-					}
-				}
-				if rsp.GetRet() != nil {
-					replica.ErrorMessage = rsp.GetRet().GetRetMsg()
-				} else {
-					replica.ErrorMessage = "empty create image response"
-				}
-				if templateID != "" && generatedReq != nil {
-					_ = UpsertReplica(ctx, templateID, generatedReq.InstanceType, replica)
-				}
-				return
-			}
-			replica.Phase = ReplicaPhaseDistributed
-			replica.CleanupRequired = false
-			replica.LastErrorPhase = ""
-			replica.ErrorMessage = ""
-			ready++
-			readyTargets = append(readyTargets, target)
-			if templateID != "" && generatedReq != nil {
-				_ = UpsertReplica(ctx, templateID, generatedReq.InstanceType, replica)
-			}
-		}()
-	}
-	wg.Wait()
-	return readyTargets, expected, ready, failed, firstErr
-}
-
-func normalizeTemplateImageRequest(req *types.CreateTemplateFromImageReq) (*types.CreateTemplateFromImageReq, error) {
-	if req == nil {
-		return nil, errors.New("request is nil")
-	}
-	if req.Request == nil || strings.TrimSpace(req.RequestID) == "" {
-		return nil, errors.New("requestID is required")
-	}
-	if strings.TrimSpace(req.SourceImageRef) == "" {
-		return nil, errors.New("source_image_ref is required")
-	}
-	if strings.HasPrefix(strings.TrimSpace(req.SourceImageRef), "-") {
-		return nil, errors.New("source_image_ref must not start with '-'")
-	}
-	if strings.TrimSpace(req.WritableLayerSize) == "" {
-		return nil, errors.New("writable_layer_size is required")
-	}
-	cloned := *req
-	exposedPorts, err := normalizeTemplateExposedPorts(req.ExposedPorts)
-	if err != nil {
-		return nil, err
-	}
-	cloned.ExposedPorts = exposedPorts
-	// Always auto-generate the template ID. Users are not allowed to set
-	// custom template IDs because the snapshot system depends on the
-	// tpl- / snap- prefix convention for storage naming and identification.
-	cloned.TemplateID = generateTemplateID()
-	if cloned.InstanceType == "" {
-		cloned.InstanceType = cubeboxv1.InstanceType_cubebox.String()
-	}
-	if cloned.NetworkType == "" {
-		cloned.NetworkType = cubeboxv1.NetworkType_tap.String()
-	}
-	return &cloned, nil
-}
-
-func buildTemplateSpecFingerprint(req *types.CreateTemplateFromImageReq, sourceImageDigest string) string {
-	type fingerprintPayload struct {
-		SourceImageDigest  string                    `json:"source_image_digest"`
-		WritableLayerSize  string                    `json:"writable_layer_size"`
-		ExposedPorts       []int32                   `json:"exposed_ports,omitempty"`
-		InstanceType       string                    `json:"instance_type"`
-		NetworkType        string                    `json:"network_type"`
-		ContainerOverrides *types.ContainerOverrides `json:"container_overrides,omitempty"`
-	}
-	payload, _ := json.Marshal(fingerprintPayload{
-		SourceImageDigest:  sourceImageDigest,
-		WritableLayerSize:  req.WritableLayerSize,
-		ExposedPorts:       req.ExposedPorts,
-		InstanceType:       req.InstanceType,
-		NetworkType:        req.NetworkType,
-		ContainerOverrides: req.ContainerOverrides,
-	})
-	sum := sha256.Sum256(payload)
-	return hex.EncodeToString(sum[:])
-}
-
-func dnsConfigOrNil(overrides *types.ContainerOverrides) *types.DNSConfig {
-	if overrides == nil {
-		return nil
-	}
-	return overrides.DnsConfig
-}
-
-func buildArtifactID(fingerprint string) string {
-	return "rfs-" + fingerprint[:24]
-}
-
-func marshalTemplateImageJobRequest(req *types.CreateTemplateFromImageReq) (string, error) {
-	if req == nil {
-		return "", errors.New("request is nil")
-	}
-	cloned := *req
-	cloned.RegistryPassword = ""
-	cloned.Request = nil
-	payload, err := json.Marshal(&cloned)
-	if err != nil {
-		return "", err
-	}
-	return string(payload), nil
-}
-
-func marshalTemplateCommitJobRequest(req *types.CreateCubeSandboxReq) (string, error) {
-	if req == nil {
-		return "", errors.New("request is nil")
-	}
-	cloned, err := cloneCreateRequest(req)
-	if err != nil {
-		return "", err
-	}
-	cloned.Request = nil
-	payload, err := json.Marshal(cloned)
-	if err != nil {
-		return "", err
-	}
-	return string(payload), nil
-}
-
-func buildCommitTemplateSpecFingerprintFromSnapshot(requestSnapshot string) string {
-	sum := sha256.Sum256([]byte(requestSnapshot))
-	return hex.EncodeToString(sum[:])
-}
-
-// buildCommitTemplateSpecFingerprint preserves the pre-merge call signature
-// used by snapshot_ops.go (it takes the unmarshaled request, hashes the
-// canonical JSON form, and returns the same value as the *FromSnapshot
-// helper would for the corresponding payload). Keeping a thin wrapper here
-// avoids touching every snapshot call site while still routing fingerprint
-// generation through a single canonical encoder.
-func buildCommitTemplateSpecFingerprint(req *types.CreateCubeSandboxReq) string {
-	payload, _ := marshalTemplateCommitJobRequest(req)
-	return buildCommitTemplateSpecFingerprintFromSnapshot(payload)
-}
-
-func getLatestTemplateImageJobByTemplateID(ctx context.Context, templateID string) (*models.TemplateImageJob, error) {
-	record := &models.TemplateImageJob{}
-	err := store.db.WithContext(ctx).Table(constants.TemplateImageJobTableName).
-		Where("template_id = ?", templateID).
-		Order("attempt_no desc, id desc").First(record).Error
-	if err != nil {
-		return nil, err
-	}
-	return record, err
-}
-
-func getTemplateImageJobByTemplateID(ctx context.Context, templateID string) (*models.TemplateImageJob, error) {
-	return getLatestTemplateImageJobByTemplateID(ctx, templateID)
-}
-
-func getActiveTemplateImageJobByTemplateID(ctx context.Context, templateID string) (*models.TemplateImageJob, error) {
-	record := &models.TemplateImageJob{}
-	err := store.db.WithContext(ctx).Table(constants.TemplateImageJobTableName).
-		Where("template_id = ? AND status IN ?", templateID, []string{JobStatusPending, JobStatusRunning}).
-		Order("attempt_no desc, id desc").First(record).Error
-	if err != nil {
-		return nil, err
-	}
-	return record, nil
-}
-
-func getTemplateImageJobByRequestID(ctx context.Context, requestID string) (*models.TemplateImageJob, error) {
-	record := &models.TemplateImageJob{}
-	err := store.db.WithContext(ctx).Table(constants.TemplateImageJobTableName).
-		Where("request_id = ?", requestID).
-		Order("id desc").First(record).Error
-	if err != nil {
-		return nil, err
-	}
-	return record, nil
-}
-
-func getActiveSnapshotJobBySandboxID(ctx context.Context, sandboxID string) (*models.TemplateImageJob, error) {
-	record := &models.TemplateImageJob{}
-	err := store.db.WithContext(ctx).Table(constants.TemplateImageJobTableName).
-		Where("sandbox_id = ? AND operation IN ? AND status IN ?", sandboxID,
-			[]string{JobOperationSnapshotCreate, JobOperationSnapshotRollback},
-			[]string{JobStatusPending, JobStatusRunning}).
-		Order("id desc").First(record).Error
-	if err != nil {
-		return nil, err
-	}
-	return record, nil
-}
-
-func getActiveSnapshotJobByResourceID(ctx context.Context, resourceID string) (*models.TemplateImageJob, error) {
-	record := &models.TemplateImageJob{}
-	err := store.db.WithContext(ctx).Table(constants.TemplateImageJobTableName).
-		Where("resource_id = ? AND operation IN ? AND status IN ?", resourceID,
-			[]string{JobOperationSnapshotCreate, JobOperationSnapshotRollback, JobOperationSnapshotDelete},
-			[]string{JobStatusPending, JobStatusRunning}).
-		Order("id desc").First(record).Error
-	if err != nil {
-		return nil, err
-	}
-	return record, nil
-}
-
-func listTemplateImageJobsByTemplateID(ctx context.Context, templateID string) ([]models.TemplateImageJob, error) {
-	var records []models.TemplateImageJob
-	err := store.db.WithContext(ctx).Table(constants.TemplateImageJobTableName).
-		Where("template_id = ?", templateID).
-		Order("attempt_no desc, id desc").Find(&records).Error
-	return records, err
-}
-
-func getRootfsArtifactByID(ctx context.Context, artifactID string) (*models.RootfsArtifact, error) {
-	record := &models.RootfsArtifact{}
-	err := store.db.WithContext(ctx).Table(constants.RootfsArtifactTableName).
-		Where("artifact_id = ?", artifactID).First(record).Error
-	if err != nil {
-		return nil, err
-	}
-	return record, err
-}
-
-func getRootfsArtifactByIDUnscoped(ctx context.Context, artifactID string) (*models.RootfsArtifact, error) {
-	record := &models.RootfsArtifact{}
-	err := store.db.WithContext(ctx).Unscoped().Table(constants.RootfsArtifactTableName).
-		Where("artifact_id = ?", artifactID).First(record).Error
-	if err != nil {
-		return nil, err
-	}
-	return record, err
-}
-
-func getRootfsArtifactByFingerprint(ctx context.Context, fingerprint string) (*models.RootfsArtifact, error) {
-	record := &models.RootfsArtifact{}
-	err := store.db.WithContext(ctx).Table(constants.RootfsArtifactTableName).
-		Where("template_spec_fingerprint = ?", fingerprint).First(record).Error
-	if err != nil {
-		return nil, err
-	}
-	return record, err
-}
-
-func getRootfsArtifactByFingerprintUnscoped(ctx context.Context, fingerprint string) (*models.RootfsArtifact, error) {
-	record := &models.RootfsArtifact{}
-	err := store.db.WithContext(ctx).Unscoped().Table(constants.RootfsArtifactTableName).
-		Where("template_spec_fingerprint = ?", fingerprint).First(record).Error
-	if err != nil {
-		return nil, err
-	}
-	return record, err
-}
-
-func updateTemplateImageJob(ctx context.Context, jobID string, values map[string]any) error {
-	values["updated_at"] = time.Now()
-	tx := store.db.WithContext(ctx).Table(constants.TemplateImageJobTableName).
-		Where("job_id = ?", jobID).Updates(values)
-	if tx.Error != nil {
-		return tx.Error
-	}
-	if tx.RowsAffected == 0 {
-		return gorm.ErrRecordNotFound
-	}
-	return nil
-}
-
-func updateRootfsArtifact(ctx context.Context, artifactID string, values map[string]any) error {
-	values["updated_at"] = time.Now()
-	tx := store.db.WithContext(ctx).Table(constants.RootfsArtifactTableName).
-		Where("artifact_id = ?", artifactID).Updates(values)
-	if tx.Error != nil {
-		return tx.Error
-	}
-	if tx.RowsAffected == 0 {
-		return gorm.ErrRecordNotFound
-	}
-	return nil
-}
-
-func jobModelToInfo(ctx context.Context, record *models.TemplateImageJob) (*types.TemplateImageJobInfo, error) {
-	info := &types.TemplateImageJobInfo{
-		JobID:                   record.JobID,
-		TemplateID:              record.TemplateID,
-		RequestID:               record.RequestID,
-		SandboxID:               record.SandboxID,
-		ResourceType:            record.ResourceType,
-		ResourceID:              record.ResourceID,
-		AttemptNo:               record.AttemptNo,
-		RetryOfJobID:            record.RetryOfJobID,
-		Operation:               record.Operation,
-		RedoMode:                record.RedoMode,
-		RedoScope:               unmarshalRedoScope(record.RedoScopeJSON),
-		ResumePhase:             record.ResumePhase,
-		ArtifactID:              record.ArtifactID,
-		TemplateSpecFingerprint: record.TemplateSpecFingerprint,
-		Status:                  record.Status,
-		Phase:                   record.Phase,
-		Progress:                record.Progress,
-		ErrorMessage:            record.ErrorMessage,
-		ExpectedNodeCount:       record.ExpectedNodeCount,
-		ReadyNodeCount:          record.ReadyNodeCount,
-		FailedNodeCount:         record.FailedNodeCount,
-		TemplateStatus:          record.TemplateStatus,
-		ArtifactStatus:          record.ArtifactStatus,
-	}
-	if record.ArtifactID != "" {
-		if artifact, err := getRootfsArtifactByID(ctx, record.ArtifactID); err == nil {
-			info.Artifact = artifactModelToInfo(artifact)
-		}
-	}
-	return info, nil
-}
-
-func artifactModelToInfo(record *models.RootfsArtifact) *types.RootfsArtifactInfo {
-	return &types.RootfsArtifactInfo{
-		ArtifactID:              record.ArtifactID,
-		TemplateSpecFingerprint: record.TemplateSpecFingerprint,
-		SourceImageRef:          record.SourceImageRef,
-		SourceImageDigest:       record.SourceImageDigest,
-		MasterNodeID:            record.MasterNodeID,
-		MasterNodeIP:            record.MasterNodeIP,
-		Ext4Path:                record.Ext4Path,
-		Ext4SHA256:              record.Ext4SHA256,
-		Ext4SizeBytes:           record.Ext4SizeBytes,
-		WritableLayerSize:       record.WritableLayerSize,
-		Status:                  record.Status,
-		LastError:               record.LastError,
-	}
-}
-
-func artifactWorkRootDir() string {
-	if value := strings.TrimSpace(os.Getenv("CUBEMASTER_ROOTFS_ARTIFACT_DIR")); value != "" {
-		return value
-	}
-	return filepath.Join(os.TempDir(), "cubemaster-rootfs-artifacts")
-}
-
-func artifactStoreRootDir() string {
-	if value := strings.TrimSpace(os.Getenv("CUBEMASTER_ROOTFS_ARTIFACT_STORE_DIR")); value != "" {
-		return value
-	}
-	return defaultArtifactStoreDir
-}
-
-func ext4FixedOverheadMiB() int64 {
-	if v := strings.TrimSpace(os.Getenv("CUBEMASTER_EXT4_FIXED_OVERHEAD_MIB")); v != "" {
-		if parsed, err := strconv.ParseInt(v, 10, 64); err == nil && parsed > 0 {
-			return parsed
-		}
-	}
-	return 256
-}
-
-func ext4OverheadPercent() int64 {
-	if v := strings.TrimSpace(os.Getenv("CUBEMASTER_EXT4_OVERHEAD_PERCENT")); v != "" {
-		if parsed, err := strconv.ParseInt(v, 10, 64); err == nil && parsed >= 1 && parsed <= 20 {
-			return parsed
-		}
-	}
-	return 10
-}
-
-func diskSpaceSafetyMargin() float64 {
-	if v := strings.TrimSpace(os.Getenv("CUBEMASTER_DISK_SPACE_SAFETY_MARGIN")); v != "" {
-		if parsed, err := strconv.ParseFloat(v, 64); err == nil && parsed >= 1.0 {
-			return parsed
-		}
-	}
-	return 1.5
-}
-
-func loopMountExt4Enabled() bool {
-	if v := strings.TrimSpace(os.Getenv("CUBEMASTER_LOOP_MOUNT_EXT4_ENABLED")); v != "" {
-		enabled, err := strconv.ParseBool(v)
-		return err == nil && enabled
-	}
-	return false
-}
-
-func artifactFallbackStoreRootDir() string {
-	return filepath.Join(os.TempDir(), fallbackArtifactStoreDir)
-}
-
-func artifactStoreDir(artifactID string) string {
-	return filepath.Join(artifactStoreRootDir(), artifactID)
-}
-
-func resolveArtifactStoreDir(ctx context.Context, artifactID string) (string, error) {
-	if configured := strings.TrimSpace(os.Getenv("CUBEMASTER_ROOTFS_ARTIFACT_STORE_DIR")); configured != "" {
-		dir := filepath.Join(configured, artifactID)
-		if err := os.MkdirAll(filepath.Dir(dir), 0o755); err != nil {
-			return "", fmt.Errorf("prepare configured artifact store root %s failed: %w", configured, err)
-		}
-		return dir, nil
-	}
-	primaryDir := artifactStoreDir(artifactID)
-	if err := os.MkdirAll(filepath.Dir(primaryDir), 0o755); err == nil {
-		return primaryDir, nil
-	} else {
-		fallbackDir := filepath.Join(artifactFallbackStoreRootDir(), artifactID)
-		if fallbackErr := os.MkdirAll(filepath.Dir(fallbackDir), 0o755); fallbackErr == nil {
-			log.G(ctx).Warnf("artifact store root %s is unavailable, fallback to %s: %v", artifactStoreRootDir(), artifactFallbackStoreRootDir(), err)
-			return fallbackDir, nil
-		} else {
-			return "", fmt.Errorf("prepare artifact store root %s failed: %w; fallback %s failed: %v", artifactStoreRootDir(), err, artifactFallbackStoreRootDir(), fallbackErr)
-		}
-	}
-}
-
-func dockerLogin(ctx context.Context, configDir, imageRef, username, password string) error {
-	registry := registryHostFromImageRef(imageRef)
-	cmd := exec.CommandContext(ctx, "docker", "--config", configDir, "login", registry, "-u", username, "--password-stdin")
-	cmd.Stdin = strings.NewReader(password)
-	output, err := cmd.CombinedOutput()
-	if err != nil {
-		return fmt.Errorf("%w: %s", err, strings.TrimSpace(string(output)))
-	}
-	return nil
-}
-
-func dockerRun(ctx context.Context, configDir string, args ...string) error {
-	_, err := dockerOutput(ctx, configDir, args...)
-	return err
-}
-
-func dockerOutput(ctx context.Context, configDir string, args ...string) ([]byte, error) {
-	cmdArgs := make([]string, 0, len(args)+2)
-	if configDir != "" {
-		cmdArgs = append(cmdArgs, "--config", configDir)
-	}
-	cmdArgs = append(cmdArgs, args...)
-	cmd := exec.CommandContext(ctx, "docker", cmdArgs...)
-	output, err := cmd.CombinedOutput()
-	if err != nil {
-		return nil, fmt.Errorf("%w: %s", err, strings.TrimSpace(string(output)))
-	}
-	return output, nil
-}
-
-func skopeoOutput(ctx context.Context, authFile string, args ...string) ([]byte, error) {
-	// skopeo expects global-ish flags such as --authfile to follow the
-	// subcommand (e.g. `skopeo inspect --authfile X docker://...`).
-	cmdArgs := make([]string, 0, len(args)+2)
-	if authFile != "" && len(args) > 0 {
-		cmdArgs = append(cmdArgs, args[0], "--authfile", authFile)
-		cmdArgs = append(cmdArgs, args[1:]...)
-	} else {
-		cmdArgs = append(cmdArgs, args...)
-	}
-	cmd := exec.CommandContext(ctx, "skopeo", cmdArgs...)
-	output, err := cmd.CombinedOutput()
-	if err != nil {
-		return nil, fmt.Errorf("%w: %s", err, strings.TrimSpace(string(output)))
-	}
-	return output, nil
-}
-
-func runCommand(ctx context.Context, dir, name string, args ...string) error {
-	cmd := exec.CommandContext(ctx, name, args...)
-	if dir != "" {
-		cmd.Dir = dir
-	}
-	output, err := cmd.CombinedOutput()
-	if err != nil {
-		return fmt.Errorf("%w: %s", err, strings.TrimSpace(string(output)))
-	}
-	return nil
-}
-
-func computeFileSHA256(path string) (string, int64, error) {
-	f, err := os.Open(path) // NOCC:Path Traversal()
-	if err != nil {
-		return "", 0, err
-	}
-	defer f.Close()
-	hasher := sha256.New()
-	// Use a 4 MiB buffer to reduce read syscall count for large files.
-	buf := make([]byte, 4*1024*1024)
-	size, err := io.CopyBuffer(hasher, f, buf)
-	if err != nil {
-		return "", 0, err
-	}
-	return hex.EncodeToString(hasher.Sum(nil)), size, nil
-}
-
-// directorySizeAndFileCount returns the total size of regular files and the count of
-// regular files in the directory tree rooted at root.  A single filepath.Walk avoids
-// a second I/O pass.
-func directorySizeAndFileCount(root string) (int64, int64, error) {
-	var totalSize, fileCount int64
-	err := filepath.Walk(root, func(path string, info os.FileInfo, walkErr error) error {
-		if walkErr != nil {
-			return walkErr
-		}
-		if info == nil || info.IsDir() {
-			return nil
-		}
-		totalSize += info.Size()
-		fileCount++
-		return nil
-	})
-	return totalSize, fileCount, err
-}
-
-// checkDiskSpace verifies that the filesystem hosting storeDir has enough free space
-// for the estimated build requirements multiplied by a configurable safety margin.
-// If storeDir does not exist yet, the function falls back to statfs-ing its parent.
-func checkDiskSpace(ctx context.Context, storeDir string, estimatedSizeBytes int64) error {
-	var stat syscall.Statfs_t
-	dir := storeDir
-	if err := syscall.Statfs(dir, &stat); err != nil {
-		if errors.Is(err, os.ErrNotExist) {
-			dir = filepath.Dir(storeDir)
-			if err2 := syscall.Statfs(dir, &stat); err2 != nil {
-				return fmt.Errorf("statfs failed for %s (and parent %s): %w", storeDir, dir, err2)
-			}
-		} else {
-			return fmt.Errorf("statfs failed for %s: %w", storeDir, err)
-		}
-	}
-	availableBytes := int64(stat.Bavail) * int64(stat.Bsize)
-	margin := diskSpaceSafetyMargin()
-	requiredBytes := int64(float64(estimatedSizeBytes) * margin)
-	if availableBytes < requiredBytes {
-		return fmt.Errorf(
-			"insufficient disk space: available=%d GiB, estimated_required=%d GiB (image_size_estimate * %.1fx safety_margin)",
-			availableBytes/(1024*1024*1024),
-			requiredBytes/(1024*1024*1024),
-			margin,
-		)
-	}
-	log.G(ctx).Infof("disk space check passed: available=%d GiB, required=%d GiB",
-		availableBytes/(1024*1024*1024), requiredBytes/(1024*1024*1024))
-	return nil
-}
-
-// isLocalFastFS returns true when the filesystem of the given path appears safe
-// for direct rootfs export rather than requiring workDir + relocate.  It is
-// conservative for known network or FUSE filesystems and falls back to the
-// parent directory when the artifact directory has not been created yet.
-func isLocalFastFS(path string) bool {
-	var stat syscall.Statfs_t
-	if err := syscall.Statfs(path, &stat); err != nil {
-		if !errors.Is(err, os.ErrNotExist) {
-			return false // cannot determine – be conservative
-		}
-		parent := filepath.Dir(path)
-		if parent == path {
-			return false
-		}
-		if err := syscall.Statfs(parent, &stat); err != nil {
-			return false
-		}
-	}
-	// Known network or FUSE filesystem magic numbers.
-	const (
-		NFS_SUPER_MAGIC  = 0x6969
-		CIFS_SUPER_MAGIC = 0xFF534D42
-		FUSE_SUPER_MAGIC = 0x65735546
-	)
-	switch stat.Type {
-	case NFS_SUPER_MAGIC, CIFS_SUPER_MAGIC, FUSE_SUPER_MAGIC:
-		return false
-	default:
-		return true
-	}
-}
-
-// canUseLoopMount checks whether the host environment supports loop-device
-// mount-based ext4 creation (Phase 2).  Returns false when CubeMaster runs
-// inside a container without CAP_SYS_ADMIN or /dev/loop-control.
-func canUseLoopMount() bool {
-	// Check CAP_SYS_ADMIN – mount(2) requires it.
-	if !hasCapability(unix.CAP_SYS_ADMIN) {
-		return false
-	}
-	// Check that /dev/loop-control exists.
-	if _, err := os.Stat("/dev/loop-control"); err != nil {
-		return false
-	}
-	// Check required commands.
-	for _, cmd := range []string{"mount", "umount", "resize2fs"} {
-		if _, err := exec.LookPath(cmd); err != nil {
-			return false
-		}
-	}
-	return true
-}
-
-// hasCapability returns true when the current process has the given capability
-// in its effective set.  On Linux this reads /proc/self/status.
-func hasCapability(cap uintptr) bool {
-	// Attempt to read CapEff from /proc/self/status.
-	data, err := os.ReadFile("/proc/self/status")
-	if err != nil {
-		return false
-	}
-	for _, line := range strings.Split(string(data), "\n") {
-		if strings.HasPrefix(line, "CapEff:") {
-			fields := strings.Fields(line)
-			if len(fields) < 2 {
-				return false
-			}
-			caps, err := strconv.ParseUint(fields[1], 16, 64)
-			if err != nil {
-				return false
-			}
-			return caps&(1<<uint(cap)) != 0
-		}
-	}
-	return false
-}
-
-// getFileBlockSize returns the actual on-disk size of a file (as reported by
-// stat(2) st_blocks * 512), which for sparse files is smaller than the apparent
-// length.
-func getFileBlockSize(path string) int64 {
-	fi, err := os.Stat(path)
-	if err != nil {
-		return 0
-	}
-	stat, ok := fi.Sys().(*syscall.Stat_t)
-	if !ok || stat == nil {
-		return fi.Size() // fallback to apparent size
-	}
-	return stat.Blocks * 512
-}
-
-const (
-	// dockerInspectSizeMultiplier expands the uncompressed on-disk image size
-	// reported by `docker image inspect` to account for rootfs data, ext4
-	// overhead, and temporary workspace.
-	dockerInspectSizeMultiplier = 4
-	// skopeoInspectSizeMultiplier expands the *compressed* layer size reported
-	// by `skopeo inspect` (LayersData[].Size). Compressed layers typically
-	// decompress ~2-3x; the extra headroom covers ext4 overhead and temporary
-	// workspace, mirroring dockerInspectSizeMultiplier on top of decompression.
-	skopeoInspectSizeMultiplier = 8
-)
-
-// skopeoLayersTotalSize sums the compressed layer blob sizes from a
-// `skopeo inspect` result. Returns 0 when no LayersData is present.
-func skopeoLayersTotalSize(info skopeoInspectImage) int64 {
-	var total int64
-	for _, layer := range info.LayersData {
-		if layer.Size > 0 {
-			total += layer.Size
-		}
-	}
-	return total
-}
-
-// estimateImageSizeFromInspect returns an approximate on-disk size for the
-// source image, used for the disk-space pre-check and Phase 2 ext4 sizing.
-//
-// The dockerless path (skopeo+umoci) must not depend on the docker binary, so
-// it derives the estimate from the compressed layer sizes captured at prepare
-// time via `skopeo inspect`. The docker path keeps using the per-image
-// cumulative Size field from `docker image inspect`.
-func estimateImageSizeFromInspect(ctx context.Context, source *resolvedSourceImage) (int64, error) {
-	if source != nil && source.useDockerless {
-		return estimateImageSizeFromSkopeo(source)
-	}
-	return estimateImageSizeFromDocker(ctx, source)
-}
-
-// estimateImageSizeFromSkopeo derives the estimate from the compressed layer
-// sizes captured by `skopeo inspect`, avoiding any docker invocation.
-func estimateImageSizeFromSkopeo(source *resolvedSourceImage) (int64, error) {
-	if source == nil || source.compressedSizeBytes <= 0 {
-		return 0, fmt.Errorf("skopeo inspect did not report any layer sizes for %s", sourceRefForLog(source))
-	}
-	return source.compressedSizeBytes * skopeoInspectSizeMultiplier, nil
-}
-
-// estimateImageSizeFromDocker extracts an approximate image size from the
-// per-image cumulative Size field in docker inspect output.
-func estimateImageSizeFromDocker(ctx context.Context, source *resolvedSourceImage) (int64, error) {
-	// Try the per-image cumulative Size field first (fast, single call).
-	out, err := dockerOutput(ctx, "", "image", "inspect", "--format", "{{.Size}}", "--", source.localRef)
-	if err != nil {
-		return 0, fmt.Errorf("docker inspect for size estimation failed: %w", err)
-	}
-	sizeBytes, err := strconv.ParseInt(strings.TrimSpace(string(out)), 10, 64)
-	if err != nil {
-		return 0, fmt.Errorf("parse docker inspect size output %q: %w", strings.TrimSpace(string(out)), err)
-	}
-	if sizeBytes <= 0 {
-		return 0, fmt.Errorf("docker image inspect reported zero or negative size (%d bytes)", sizeBytes)
-	}
-	// RootFS data plus writable layer overhead typically expands 3-4x from the
-	// on-disk layer size. Use a conservative multiplier for the disk-space
-	// pre-check.
-	return sizeBytes * dockerInspectSizeMultiplier, nil
-}
-
-func sourceRefForLog(source *resolvedSourceImage) string {
-	if source == nil {
-		return "<nil>"
-	}
-	return source.localRef
-}
-
-// createExt4ImageStreaming uses loop-mount to stream docker export directly into
-// an ext4 image, avoiding any intermediate rootfs directory on disk (Phase 2).
-// Falls back to Phase 1 when prerequisites are not met.
-// estimatedSizeBytes should be obtained from estimateImageSizeFromInspect.
-func createExt4ImageStreaming(ctx context.Context, source *resolvedSourceImage, workDir, ext4Path string, estimatedSizeBytes int64) error {
-	if !canUseLoopMount() {
-		return fmt.Errorf("loop mount not available")
-	}
-
-	// 2. Create empty ext4 image using the caller-provided size estimate.
-	if err := runCommand(ctx, "", "truncate", "-s", strconv.FormatInt(estimatedSizeBytes, 10), ext4Path); err != nil {
-		return fmt.Errorf("truncate ext4 image for streaming: %w", err)
-	}
-	if err := runCommand(ctx, "", "mkfs.ext4", "-F", ext4Path); err != nil {
-		return fmt.Errorf("mkfs.ext4 for streaming: %w", err)
-	}
-
-	// 3. Mount the ext4 image via loop device.
-	mountPoint := filepath.Join(workDir, "ext4-mnt")
-	if err := os.MkdirAll(mountPoint, 0o700); err != nil {
-		return fmt.Errorf("create mount point: %w", err)
-	}
-
-	// Use context.Background() for cleanup so it runs even after request cancellation.
-	cleanupCtx := context.Background()
-	var unmountOnce sync.Once
-	var detachOnce sync.Once
-	cleanup := func() {
-		unmountOnce.Do(func() {
-			_ = runCommand(cleanupCtx, "", "umount", "--", mountPoint)
-		})
-		if err := os.RemoveAll(mountPoint); err != nil {
-			log.G(ctx).Warnf("cleanup mount point %s failed: %v", mountPoint, err)
-		}
-	}
-	defer cleanup()
-
-	// Allocate a free loop device and mount.
-	loopOut, err := exec.CommandContext(ctx, "losetup", "--find", "--show", ext4Path).CombinedOutput()
-	if err != nil {
-		return fmt.Errorf("losetup --find --show %s failed: %w: %s", ext4Path, err, string(loopOut))
-	}
-	loopDevice := strings.TrimSpace(string(loopOut))
-	detachLoop := func() {
-		detachOnce.Do(func() {
-			_ = runCommand(cleanupCtx, "", "losetup", "--detach", "--", loopDevice)
-		})
-	}
-	defer detachLoop()
-
-	if err := runCommand(ctx, "", "mount", "-o", "nosuid,noexec,nodev,noatime", "--", loopDevice, mountPoint); err != nil {
-		detachLoop() // explicit detach on mount failure (defer will be a no-op via sync.Once)
-		return fmt.Errorf("mount loop device %s: %w", loopDevice, err)
-	}
-
-	// 4. Create container and stream export directly into the mounted ext4.
-	containerIDBytes, err := dockerOutput(ctx, "", "create", "--", source.localRef)
-	if err != nil {
-		return fmt.Errorf("docker create for streaming: %w", err)
-	}
-	containerID := strings.TrimSpace(string(containerIDBytes))
-	defer func() {
-		_ = dockerRun(cleanupCtx, "", "rm", "-f", containerID)
-	}()
-
-	if err := pipeExportToDir(ctx, containerID, mountPoint); err != nil {
-		return fmt.Errorf("pipe export to mount point: %w", err)
-	}
-
-	// 5. Unmount (via cleanup).
-	cleanup()
-
-	// 6. Shrink the ext4 filesystem to minimum size (best-effort).
-	if err := runCommand(cleanupCtx, "", "resize2fs", "-M", ext4Path); err != nil {
-		log.G(ctx).Warnf("resize2fs -M failed (best-effort, using original size): %v", err)
-	}
-
-	// 7. Truncate to actual block size.
-	finalSize := getFileBlockSize(ext4Path)
-	if finalSize > 0 && finalSize < estimatedSizeBytes {
-		if err := runCommand(cleanupCtx, "", "truncate", "-s", strconv.FormatInt(finalSize, 10), ext4Path); err != nil {
-			log.G(ctx).Warnf("truncate to final size %d failed: %v", finalSize, err)
-		}
-	}
-
-	return nil
-}
-
-// pipeExportToDir streams the docker export of a container directly into a target
-// directory via tar -xf -.
-func pipeExportToDir(ctx context.Context, containerID, destDir string) error {
-	exportCmd := exec.CommandContext(ctx, "docker", "export", containerID)
-	tarCmd := exec.CommandContext(ctx, "tar", "-xf", "-", "-C", destDir)
-
-	pipe, err := exportCmd.StdoutPipe()
-	if err != nil {
-		return fmt.Errorf("create pipe from docker export to tar: %w", err)
-	}
-	tarCmd.Stdin = pipe
-
-	// Best-effort: increase pipe buffer to 1 MiB.
-	if f, ok := pipe.(*os.File); ok {
-		_, _ = unix.FcntlInt(f.Fd(), unix.F_SETPIPE_SZ, 1<<20)
-	}
-
-	var exportErrBuf, tarErrBuf bytes.Buffer
-	exportCmd.Stderr = &exportErrBuf
-	tarCmd.Stderr = &tarErrBuf
-
-	if err := tarCmd.Start(); err != nil {
-		return fmt.Errorf("start tar extract: %w", err)
-	}
-	if err := exportCmd.Start(); err != nil {
-		tarCmd.Process.Kill()
-		_ = tarCmd.Wait()
-		return fmt.Errorf("start docker export: %w", err)
-	}
-
-	exportWaitErr := exportCmd.Wait()
-	tarWaitErr := tarCmd.Wait()
-
-	if exportWaitErr != nil {
-		return fmt.Errorf("docker export %s failed: %w (stderr: %s)", containerID, exportWaitErr, exportErrBuf.String())
-	}
-	if tarWaitErr != nil {
-		return fmt.Errorf("extract tar to %s failed: %w (stderr: %s)", destDir, tarWaitErr, tarErrBuf.String())
-	}
-	return nil
-}
-
-func firstNonEmptyDigest(info dockerInspectImage) string {
-	if len(info.RepoDigests) > 0 && info.RepoDigests[0] != "" {
-		rd := info.RepoDigests[0]
-		// RepoDigests entries are canonical references of the form
-		// "name@sha256:...". We only want the digest portion so that
-		// callers can compose "ref@digest" without producing
-		// "name:tag@name@sha256:..." style duplication.
-		if at := strings.Index(rd, "@"); at >= 0 && at+1 < len(rd) {
-			return rd[at+1:]
-		}
-		return rd
-	}
-	return info.ID
-}
-
-func envListToKeyValues(envs []string) []*types.KeyValue {
-	if len(envs) == 0 {
-		return nil
-	}
-	out := make([]*types.KeyValue, 0, len(envs))
-	for _, env := range envs {
-		parts := strings.SplitN(env, "=", 2)
-		kv := &types.KeyValue{Key: parts[0]}
-		if len(parts) == 2 {
-			kv.Value = parts[1]
-		}
-		out = append(out, kv)
-	}
-	sort.SliceStable(out, func(i, j int) bool {
-		return out[i].Key < out[j].Key
-	})
-	return out
-}
-
-func defaultRLimit(overrides *types.ContainerOverrides) *types.RLimit {
-	if overrides != nil && overrides.RLimit != nil {
-		return overrides.RLimit
-	}
-	return &types.RLimit{NoFile: 1000000}
-}
-
-func probeOrNil(overrides *types.ContainerOverrides) *types.Probe {
-	if overrides == nil {
-		return nil
-	}
-	return overrides.Probe
-}
-
-func buildDownloadURL(baseURL, artifactID, token string) string {
-	trimmed := strings.TrimRight(normalizeBaseURL(baseURL), "/")
-	if trimmed == "" {
-		trimmed = "http://" + artifactRootHostHint()
-	}
-	u, err := url.Parse(trimmed + "/cube/template/artifact/download")
-	if err != nil {
-		return trimmed
-	}
-	query := u.Query()
-	query.Set("artifact_id", artifactID)
-	query.Set("token", token)
-	u.RawQuery = query.Encode()
-	return u.String()
-}
-
-func normalizeBaseURL(baseURL string) string {
-	trimmed := strings.TrimSpace(baseURL)
-	if trimmed == "" {
-		return ""
-	}
-	if strings.HasPrefix(trimmed, "http://") || strings.HasPrefix(trimmed, "https://") {
-		return trimmed
-	}
-	return "http://" + trimmed
-}
-
-func artifactRootHostHint() string {
-	host, err := os.Hostname()
-	if err != nil || host == "" {
-		return "127.0.0.1"
-	}
-	return host
-}
-
-func registryHostFromImageRef(imageRef string) string {
-	imageRef = strings.TrimPrefix(imageRef, "docker://")
-	parts := strings.Split(imageRef, "/")
-	if len(parts) == 0 {
-		return "docker.io"
-	}
-	first := parts[0]
-	if strings.Contains(first, ".") || strings.Contains(first, ":") || first == "localhost" {
-		return first
-	}
-	return "docker.io"
-}
-
-func quantityToGi(value string) (int64, error) {
-	v := strings.TrimSpace(strings.ToLower(value))
-	switch {
-	case strings.HasSuffix(v, "gi"):
-		return strconv.ParseInt(strings.TrimSuffix(v, "gi"), 10, 64)
-	case strings.HasSuffix(v, "g"):
-		return strconv.ParseInt(strings.TrimSuffix(v, "g"), 10, 64)
-	case strings.HasSuffix(v, "mi"):
-		mi, err := strconv.ParseInt(strings.TrimSuffix(v, "mi"), 10, 64)
-		if err != nil {
-			return 0, err
-		}
-		if mi%1024 == 0 {
-			return mi / 1024, nil
-		}
-		return mi/1024 + 1, nil
-	default:
-		return strconv.ParseInt(v, 10, 64)
-	}
-}
-
-func containsString(items []string, target string) bool {
-	for _, item := range items {
-		if item == target {
-			return true
-		}
-	}
-	return false
-}
-
-func normalizeTemplateExposedPorts(ports []int32) ([]int32, error) {
-	if len(ports) == 0 {
-		return nil, nil
-	}
-	uniq := make(map[int32]struct{}, len(ports))
-	normalized := make([]int32, 0, len(ports))
-	for _, port := range ports {
-		if port <= 0 || port > 65535 {
-			return nil, fmt.Errorf("invalid exposed port %d", port)
-		}
-		if _, exists := uniq[port]; exists {
-			continue
-		}
-		uniq[port] = struct{}{}
-		normalized = append(normalized, port)
-	}
-	sort.Slice(normalized, func(i, j int) bool {
-		return normalized[i] < normalized[j]
-	})
-	if countCustomTemplateExposedPorts(normalized) > 3 {
-		return nil, fmt.Errorf("at most 3 custom exposed ports are supported")
-	}
-	return normalized, nil
-}
-
-func countCustomTemplateExposedPorts(ports []int32) int {
-	reserved := defaultTemplateExposedPorts()
-	count := 0
-	for _, port := range ports {
-		if _, ok := reserved[port]; ok {
-			continue
-		}
-		count++
-	}
-	return count
-}
-
-func defaultTemplateExposedPorts() map[int32]struct{} {
-	return map[int32]struct{}{
-		49983: {},
-	}
-}
-
-func formatExposedPortsAnnotation(ports []int32) string {
-	if len(ports) == 0 {
-		return ""
-	}
-	values := make([]string, 0, len(ports))
-	for _, port := range ports {
-		values = append(values, strconv.FormatInt(int64(port), 10))
-	}
-	return strings.Join(values, ":")
-}
-
-func errorString(err error) string {
-	if err == nil {
-		return ""
-	}
-	return err.Error()
-}
-
-func detachTemplateImageJobContext(ctx context.Context, fields map[string]any) context.Context {
-	detached := context.Background()
-	if rt := CubeLog.GetTraceInfo(ctx); rt != nil {
-		detached = CubeLog.WithRequestTrace(detached, rt.DeepCopy())
-	}
-	return log.WithLogger(detached, log.G(ctx).WithFields(fields))
-}
-
-func ensureArtifactBuildPreflight(ctx context.Context) error {
-	requiredCommands := []string{"mkfs.ext4", "truncate", "cp"}
-	if hasDockerlessRootfsExportTools() {
-		requiredCommands = append(requiredCommands, "skopeo", "umoci")
-	} else {
-		requiredCommands = append(requiredCommands, "docker", "tar")
-	}
-	for _, cmd := range requiredCommands {
-		if _, err := executableLookPath(cmd); err != nil {
-			return fmt.Errorf("required command %q is not available on cubemaster node", cmd)
-		}
-	}
-	output, err := exec.CommandContext(ctx, "mkfs.ext4", "-h").CombinedOutput()
-	helpText := string(output)
-	if err != nil && helpText == "" {
-		return fmt.Errorf("failed to probe mkfs.ext4 help output: %w", err)
-	}
-	if !strings.Contains(helpText, "-d") {
-		return fmt.Errorf("mkfs.ext4 on cubemaster node does not appear to support the -d option required for rootfs image creation")
-	}
-	return nil
-}
-
-func relocateRootfsToArtifactStore(ctx context.Context, srcRootfsDir, dstRootfsDir string) error {
-	if err := os.RemoveAll(dstRootfsDir); err != nil { // NOCC:Path Traversal()
-		return err
-	}
-	if err := os.MkdirAll(filepath.Dir(dstRootfsDir), 0o755); err != nil {
-		return err
-	}
-	if err := os.Rename(srcRootfsDir, dstRootfsDir); err == nil {
-		return nil
-	} else if !isCrossDeviceRenameErr(err) {
-		return err
-	}
-	if err := runCommand(ctx, "", "cp", "-a", srcRootfsDir, dstRootfsDir); err != nil {
-		return fmt.Errorf("copy rootfs to artifact store failed: %w", err)
-	}
-	return os.RemoveAll(srcRootfsDir) // NOCC:Path Traversal()
-}
-
-func isCrossDeviceRenameErr(err error) bool {
-	var linkErr *os.LinkError
-	if errors.As(err, &linkErr) {
-		return errors.Is(linkErr.Err, syscall.EXDEV)
-	}
-	return errors.Is(err, syscall.EXDEV)
-}
-
-func cleanupFailedRootfsArtifact(ctx context.Context, artifact *models.RootfsArtifact, instanceType string) error {
-	if artifact == nil {
-		return nil
-	}
-	var cleanupErr error
-	if err := cleanupDistributedArtifact(ctx, artifact.ArtifactID, instanceType); err != nil {
-		cleanupErr = errors.Join(cleanupErr, err)
-	}
-	if err := cleanupLocalRootfsArtifact(artifact.ArtifactID, artifact.Ext4Path); err != nil {
-		cleanupErr = errors.Join(cleanupErr, err)
-	}
-	if cleanupErr == nil {
-		if err := deleteRootfsArtifactRecord(ctx, artifact.ArtifactID); err == nil {
-			return nil
-		} else {
-			cleanupErr = errors.Join(cleanupErr, err)
-		}
-	}
-	updateErr := updateRootfsArtifact(ctx, artifact.ArtifactID, map[string]any{
-		"status":     ArtifactStatusFailed,
-		"last_error": fmt.Sprintf("artifact cleanup incomplete: %v", cleanupErr),
-	})
-	return errors.Join(cleanupErr, updateErr)
-}
-
-func cleanupLocalRootfsArtifact(artifactID, ext4Path string) error {
-	if ext4Path == "" {
-		return nil
-	}
-	if dir, ok := managedArtifactDir(artifactID, ext4Path); ok {
-		return os.RemoveAll(dir) // NOCC:Path Traversal()
-	}
-	if err := os.Remove(ext4Path); err != nil && !errors.Is(err, os.ErrNotExist) { // NOCC:Path Traversal()
-		return err
-	}
-	return nil
-}
-
-func managedArtifactDir(artifactID, ext4Path string) (string, bool) {
-	if strings.TrimSpace(artifactID) == "" || strings.TrimSpace(ext4Path) == "" {
-		return "", false
-	}
-	dir := filepath.Clean(filepath.Dir(ext4Path))
-	if filepath.Base(dir) != artifactID {
-		return "", false
-	}
-	roots := []string{artifactWorkRootDir(), artifactStoreRootDir()}
-	if strings.TrimSpace(os.Getenv("CUBEMASTER_ROOTFS_ARTIFACT_STORE_DIR")) == "" {
-		roots = append(roots, artifactFallbackStoreRootDir())
-	}
-	for _, root := range roots {
-		rel, err := filepath.Rel(filepath.Clean(root), dir)
-		if err != nil {
-			continue
-		}
-		if rel == artifactID {
-			return dir, true
-		}
-	}
-	return "", false
 }
